now have a python export of precomp
