--- wisdem/precomp/test_precomp.py
@@ -1,1616 +1,0 @@
-import unittest
-
-import numpy as np
-import numpy.testing as npt
-
-import properties as prop
-from _precomp import precomp as _precomp
-
-class TestPreCompProperties(unittest.TestCase):
-    def test_tw_rate(self):
-
-        # From NREL 5 MW
-        myr = np.array([ 1.5,         3.62068966,  5.74137931,  7.86206897,  9.98275862, 12.10344828,
-                         14.22413793, 16.34482759, 18.46551724, 20.5862069 , 22.70689655, 24.82758621,
-                         26.94827586, 29.06896552, 31.18965517, 33.31034483, 35.43103448, 37.55172414,
-                         39.67241379, 41.79310345, 43.9137931 , 46.03448276, 48.15517241, 50.27586207,
-                         52.39655172, 54.51724138, 56.63793103, 58.75862069, 60.87931034, 63.        ])
-        th0 = np.zeros(myr.shape)
-        th_prime_fort = _precomp.tw_rate(myr, th0)
-        #th_prime_fort = th0
-        th_prime_py = prop.tw_rate(myr, th0)
-        npt.assert_almost_equal(th_prime_fort, th_prime_py)
-
-        th0 = np.linspace(10., -10., myr.size)
-        th_prime_fort = _precomp.tw_rate(myr, th0)
-        #th_prime_fort = -0.32520325*np.ones(th0.shape)
-        th_prime_py = prop.tw_rate(myr, th0)
-        npt.assert_almost_equal(th_prime_fort, th_prime_py)
-
-    def test_properties(self):
-        # From NREL 5 MW station 0
-        chord = 3.542
-        theta = th_prime = 0.0
-        le_loc = 0.5
-        xnode = np.array([0.00000000e+00, 2.51728808e-04, 1.00666176e-03, 2.26403871e-03,
-                          4.02259358e-03, 6.28055566e-03, 9.03565137e-03, 1.22851066e-02,
-                          1.60256493e-02, 2.02535132e-02, 2.49644411e-02, 3.01536896e-02,
-                          3.58160335e-02, 4.19457713e-02, 4.85367309e-02, 5.55822757e-02,
-                          6.30753115e-02, 7.10082934e-02, 7.93732336e-02, 8.81617093e-02,
-                          9.73648712e-02, 1.06973453e-01, 1.16977778e-01, 1.27367775e-01,
-                          1.38132981e-01, 1.49262556e-01, 1.60745294e-01, 1.72569633e-01,
-                          1.84723666e-01, 1.97195156e-01, 2.09971545e-01, 2.23039968e-01,
-                          2.36387266e-01, 2.50000000e-01, 2.63864463e-01, 2.77966694e-01,
-                          2.92292493e-01, 3.06827437e-01, 3.21556889e-01, 3.36466018e-01,
-                          3.51539812e-01, 3.66763093e-01, 3.82120532e-01, 3.97596666e-01,
-                          4.13175911e-01, 4.28842581e-01, 4.44580900e-01, 4.60375022e-01,
-                          4.76209042e-01, 4.92067018e-01, 5.07932982e-01, 5.23790958e-01,
-                          5.39624978e-01, 5.55419100e-01, 5.71157419e-01, 5.86824089e-01,
-                          6.02403334e-01, 6.17879468e-01, 6.33236907e-01, 6.48460188e-01,
-                          6.63533982e-01, 6.78443111e-01, 6.93172563e-01, 7.07707507e-01,
-                          7.22033306e-01, 7.36135537e-01, 7.50000000e-01, 7.63612734e-01,
-                          7.76960032e-01, 7.90028455e-01, 8.02804844e-01, 8.15276334e-01,
-                          8.27430367e-01, 8.39254706e-01, 8.50737444e-01, 8.61867019e-01,
-                          8.72632225e-01, 8.83022222e-01, 8.93026547e-01, 9.02635129e-01,
-                          9.11838291e-01, 9.20626766e-01, 9.28991707e-01, 9.36924689e-01,
-                          9.44417724e-01, 9.51463269e-01, 9.58054229e-01, 9.64183967e-01,
-                          9.69846310e-01, 9.75035559e-01, 9.79746487e-01, 9.83974351e-01,
-                          9.87714893e-01, 9.90964349e-01, 9.93719444e-01, 9.95977406e-01,
-                          9.97735961e-01, 9.98993338e-01, 9.99748271e-01, 1.00000000e+00,
-                          9.99748271e-01, 9.98993338e-01, 9.97735961e-01, 9.95977406e-01,
-                          9.93719444e-01, 9.90964349e-01, 9.87714893e-01, 9.83974351e-01,
-                          9.79746487e-01, 9.75035559e-01, 9.69846310e-01, 9.64183967e-01,
-                          9.58054229e-01, 9.51463269e-01, 9.44417724e-01, 9.36924689e-01,
-                          9.28991707e-01, 9.20626766e-01, 9.11838291e-01, 9.02635129e-01,
-                          8.93026547e-01, 8.83022222e-01, 8.72632225e-01, 8.61867019e-01,
-                          8.50737444e-01, 8.39254706e-01, 8.27430367e-01, 8.15276334e-01,
-                          8.02804844e-01, 7.90028455e-01, 7.76960032e-01, 7.63612734e-01,
-                          7.50000000e-01, 7.36135537e-01, 7.22033306e-01, 7.07707507e-01,
-                          6.93172563e-01, 6.78443111e-01, 6.63533982e-01, 6.48460188e-01,
-                          6.33236907e-01, 6.17879468e-01, 6.02403334e-01, 5.86824089e-01,
-                          5.71157419e-01, 5.55419100e-01, 5.39624978e-01, 5.23790958e-01,
-                          5.07932982e-01, 4.92067018e-01, 4.76209042e-01, 4.60375022e-01,
-                          4.44580900e-01, 4.28842581e-01, 4.13175911e-01, 3.97596666e-01,
-                          3.82120532e-01, 3.66763093e-01, 3.51539812e-01, 3.36466018e-01,
-                          3.21556889e-01, 3.06827437e-01, 2.92292493e-01, 2.77966694e-01,
-                          2.63864463e-01, 2.50000000e-01, 2.36387266e-01, 2.23039968e-01,
-                          2.09971545e-01, 1.97195156e-01, 1.84723666e-01, 1.72569633e-01,
-                          1.60745294e-01, 1.49262556e-01, 1.38132981e-01, 1.27367775e-01,
-                          1.16977778e-01, 1.06973453e-01, 9.73648712e-02, 8.81617093e-02,
-                          7.93732336e-02, 7.10082934e-02, 6.30753115e-02, 5.55822757e-02,
-                          4.85367309e-02, 4.19457713e-02, 3.58160335e-02, 3.01536896e-02,
-                          2.49644411e-02, 2.02535132e-02, 1.60256493e-02, 1.22851066e-02,
-                          9.03565137e-03, 6.28055566e-03, 4.02259358e-03, 2.26403871e-03,
-                          1.00666176e-03, 2.51728808e-04])
-        ynode = np.array([ 0.         , 0.01462169 , 0.03093556 , 0.04686826 , 0.06258357 , 0.07861745,
-                           0.09405252 , 0.10967754 , 0.12517467 , 0.14030205 , 0.15556546 , 0.17061296,
-                           0.18530152 , 0.19996301 , 0.21449599 , 0.22874252 , 0.24262855 , 0.25634017,
-                           0.26983858 , 0.28309181 , 0.29606886 , 0.30870461 , 0.32101182 , 0.33297108,
-                           0.34460585 , 0.35590175 , 0.36684564 , 0.37742782 , 0.38764095 , 0.39746876,
-                           0.40687689 , 0.4158685  , 0.4244587  , 0.43265564 , 0.44035434 , 0.44758788,
-                           0.45439201 , 0.4607208  , 0.4666015  , 0.4720264  , 0.47698672 , 0.48148686,
-                           0.48552288 , 0.48903208 , 0.49198712 , 0.4944485  , 0.49644181 , 0.49797497,
-                           0.49903068 , 0.49956586 , 0.49951122 , 0.49896566 , 0.49795529 , 0.496483,
-                           0.49454021 , 0.49197826 , 0.488932   , 0.48544118 , 0.4815031  , 0.47706314,
-                           0.47204391 , 0.46659337 , 0.46072373 , 0.45444908 , 0.44755198 , 0.44023838,
-                           0.4325517  , 0.42450843 , 0.41583676 , 0.40678988 , 0.39741734 , 0.38773107,
-                           0.37741607 , 0.36678968 , 0.35588415 , 0.34465455 , 0.33287266 , 0.32086115,
-                           0.30864941 , 0.29605157 , 0.28300335 , 0.26980349 , 0.25650823 , 0.24258332,
-                           0.22849786 , 0.21437412 , 0.20008925 , 0.18538938 , 0.17029503 , 0.15488099,
-                           0.13980699 , 0.12467987 , 0.10951799 , 0.09361538 , 0.07689355 , 0.06143269,
-                           0.04835402 , 0.03087902 , 0.0122301  , 0.00369928, -0.01428549, -0.03074717,
-                           -0.04732177, -0.06375087, -0.07907204, -0.09463611, -0.11041152, -0.12590768,
-                          -0.14098449, -0.15620072, -0.17142561, -0.18610004, -0.20064589, -0.21515702,
-                          -0.2295169 , -0.24337121, -0.2570841 , -0.27063306, -0.28390427, -0.29672722,
-                          -0.30934207, -0.32172203, -0.33375775, -0.34532288, -0.35662667, -0.36763219,
-                          -0.37826137, -0.38840314, -0.39819124, -0.40761009, -0.41665301, -0.4252099,
-                          -0.43334852, -0.4410498 , -0.44831686, -0.45518375, -0.46154504, -0.46740234,
-                          -0.47283049, -0.47778693, -0.48225506, -0.48627201, -0.48977107, -0.49274869,
-                          -0.49525282, -0.49725252, -0.49875597, -0.49976917, -0.50027517, -0.50029574,
-                          -0.49981064, -0.4988094 , -0.49728292, -0.49526513, -0.49277202, -0.48977413,
-                          -0.48627141, -0.482278  , -0.47780307, -0.47284984, -0.46742386, -0.46153196,
-                          -0.45518172, -0.4483774 , -0.44107921, -0.43334623, -0.4251908 , -0.41662661,
-                          -0.40766583, -0.39826616, -0.38843777, -0.37820676, -0.36765976, -0.35672833,
-                          -0.34538119, -0.33373416, -0.32178271, -0.30940667, -0.29682167, -0.28386015,
-                          -0.27065183, -0.25715717, -0.24348009, -0.22943138, -0.21522871, -0.20084706,
-                          -0.18616117, -0.17135669, -0.15630674, -0.14115807, -0.12588878, -0.11053819,
-                          -0.09491395, -0.0792819 , -0.06370407, -0.04766179, -0.03168923, -0.01532304])
-        
-        E1 = np.array([3440000000.0, 41800000000.0, 27700000000.0, 13600000000.0, 200000000000.0, 205000000000.0, 118000000000.0, 98460000.0, 114500000000.0, 1000000.0, 4560000.0])
-        E2 = np.array([3440000000.0, 14000000000.0, 13650000000.0, 13300000000.0, 200000000000.0, 205000000000.0, 118000000000.0, 98460000.0, 8390000000.0, 1000000.0, 4560000.0])
-        G12 = np.array([1323000000.0, 2630000000.0, 7200000000.0, 11800000000.0, 79300000000.0, 80000000000.0, 47600000000.0, 98460000.0, 5990000000.0, 384615.3846153846, 1520000.0])
-        nu12 = np.array([0.3, 0.28, 0.39, 0.49, 0.3, 0.3, 0.3, 0.3, 0.27, 0.3, 0.49])
-        rho = np.array([1235.0, 1920.0, 1850.0, 1780.0, 7800.0, 7850.0, 7200.0, 200.0, 1220.0, 1150.0, 1100.0])
-        locU = np.array([0.        , 0.62506011, 0.81300099, 1.        ])
-        n_laminaU = np.array([4., 5., 4.])
-        n_pliesU = np.array([1., 1., 1., 1., 1., 1., 1., 1., 1., 1., 1., 1., 1.])
-        tU = np.array([5.00e-05, 2.82e-03, 5.17e-02, 2.82e-03, 5.00e-05, 2.82e-03, 5.17e-02, 4.70e-04, 2.82e-03, 5.00e-05, 2.82e-03, 5.17e-02, 2.82e-03])
-        thetaU = np.array([0., 0., 0., 0., 0., 0., 0., 0., 0., 0., 0., 0., 0.])
-        mat_idxU = np.array([1, 3, 3, 3, 1, 3, 3, 9, 3, 1, 3, 3, 3], dtype=np.int_)
-        locL = np.array([0.        , 0.18705037, 0.37497357, 1.        ])
-        n_laminaL = np.array([4., 5., 4.])
-        n_pliesL = np.array([1., 1., 1., 1., 1., 1., 1., 1., 1., 1., 1., 1., 1.])
-        tL = np.array([5.00e-05, 2.82e-03, 5.17e-02, 2.82e-03, 5.00e-05, 2.82e-03, 5.17e-02, 4.70e-04, 2.82e-03, 5.00e-05, 2.82e-03, 5.17e-02, 2.82e-03])
-        thetaL = np.array([0., 0., 0., 0., 0., 0., 0., 0., 0., 0., 0., 0., 0.])
-        mat_idxL = np.array([1, 3, 3, 3, 1, 3, 3, 9, 3, 1, 3, 3, 3], dtype=np.int_)
-        nwebs=2
-        locW = np.array([0.64505453, 0.79653854])
-        n_laminaW = np.array([3., 3.])
-        n_pliesW = np.array([1., 1., 1., 1., 1., 1.])
-        tW = np.array([0.002, 0.04 , 0.002, 0.002, 0.04 , 0.002])
-        thetaW = np.array([0., 0., 0., 0., 0., 0.])
-        mat_idxW = np.array([4, 8, 4, 4, 8, 4], dtype=np.int_)
-
-
-        (eifbar,eilbar,gjbar,eabar,eiflbar,
-         sfbar,slbar,sftbar,sltbar,satbar,
-         z_sc,y_sc,ztc_ref,ytc_ref,
-         mass,area,iflap_eta,ilag_zeta,tw_iner,
-         zcm_ref,ycm_ref) = _precomp.properties(
-            chord,
-            theta,
-            th_prime,
-            le_loc,
-            xnode,
-            ynode,
-            E1,
-            E2,
-            G12,
-            nu12,
-            rho,
-            locU,
-            n_laminaU,
-            n_pliesU,
-            tU,
-            thetaU,
-            mat_idxU,
-            locL,
-            n_laminaL,
-            n_pliesL,
-            tL,
-            thetaL,
-            mat_idxL,
-            nwebs,
-            locW,
-            n_laminaW,
-            n_pliesW,
-            tW,
-            thetaW,
-            mat_idxW,
-        )
-        results_fort = (eifbar,eilbar,gjbar,eabar,eiflbar,
-         sfbar,slbar,sftbar,sltbar,satbar,
-         z_sc,y_sc,ztc_ref,ytc_ref,
-         mass,area,iflap_eta,ilag_zeta,tw_iner,
-         zcm_ref,ycm_ref)
-        #results_fort = (27308516755.423786, 27110066507.9137, 13952243195.444218, 18118671160.42156, 92074585.33274896, -15460.740430957281, -280289306.6835994, -0.0, 0.0, 0.0, -0.0013892862524830067, -6.888959056499682e-05, -0.0013884329481598547, 0.015400749667299764, 1276.1536419953366, 17.366065786576875, 1872.3066046676142, 1851.0336153527403, -2.5971291329272783, -0.0013858771158858443, 0.056855449043146725)
-        print(results_fort)
-        
-        (eifbar, eilbar, gjbar, eabar, eiflbar,
-         sfbar, slbar, sftbar, sltbar, satbar,
-         z_sc, y_sc, ztc_ref, ytc_ref,
-         mass, area, iflap_eta, ilag_zeta, tw_iner,
-         zcm_ref, ycm_ref) = prop.properties(chord,
-                                     theta,
-                                     th_prime,
-                                     le_loc,
-                                     xnode,
-                                     ynode,
-                                     E1,
-                                     E2,
-                                     G12,
-                                     nu12,
-                                     rho,
-                                     locU,
-                                     n_laminaU,
-                                     n_pliesU,
-                                     tU,
-                                     thetaU,
-                                     mat_idxU,
-                                     locL,
-                                     n_laminaL,
-                                     n_pliesL,
-                                     tL,
-                                     thetaL,
-                                     mat_idxL,
-                                     nwebs,
-                                     locW,
-                                     n_laminaW,
-                                     n_pliesW,
-                                     tW,
-                                     thetaW,
-                                     mat_idxW,
-                                     )
-        results_py = (eifbar,eilbar,gjbar,eabar,eiflbar,
-         sfbar,slbar,sftbar,sltbar,satbar,
-         z_sc,y_sc,ztc_ref,ytc_ref,
-         mass,area,iflap_eta,ilag_zeta,tw_iner,
-         zcm_ref,ycm_ref)
-        print(results_py)
-        npt.assert_almost_equal(results_fort, results_py)
-
-        '''
-            
-0 3.542 0.0 0.0 0.5 [0.00000000e+00 2.51728808e-04 1.00666176e-03 2.26403871e-03
- 4.02259358e-03 6.28055566e-03 9.03565137e-03 1.22851066e-02
- 1.60256493e-02 2.02535132e-02 2.49644411e-02 3.01536896e-02
- 3.58160335e-02 4.19457713e-02 4.85367309e-02 5.55822757e-02
- 6.30753115e-02 7.10082934e-02 7.93732336e-02 8.81617093e-02
- 9.73648712e-02 1.06973453e-01 1.16977778e-01 1.27367775e-01
- 1.38132981e-01 1.49262556e-01 1.60745294e-01 1.72569633e-01
- 1.84723666e-01 1.97195156e-01 2.09971545e-01 2.23039968e-01
- 2.36387266e-01 2.50000000e-01 2.63864463e-01 2.77966694e-01
- 2.92292493e-01 3.06827437e-01 3.21556889e-01 3.36466018e-01
- 3.51539812e-01 3.66763093e-01 3.82120532e-01 3.97596666e-01
- 4.13175911e-01 4.28842581e-01 4.44580900e-01 4.60375022e-01
- 4.76209042e-01 4.92067018e-01 5.07932982e-01 5.23790958e-01
- 5.39624978e-01 5.55419100e-01 5.71157419e-01 5.86824089e-01
- 6.02403334e-01 6.17879468e-01 6.33236907e-01 6.48460188e-01
- 6.63533982e-01 6.78443111e-01 6.93172563e-01 7.07707507e-01
- 7.22033306e-01 7.36135537e-01 7.50000000e-01 7.63612734e-01
- 7.76960032e-01 7.90028455e-01 8.02804844e-01 8.15276334e-01
- 8.27430367e-01 8.39254706e-01 8.50737444e-01 8.61867019e-01
- 8.72632225e-01 8.83022222e-01 8.93026547e-01 9.02635129e-01
- 9.11838291e-01 9.20626766e-01 9.28991707e-01 9.36924689e-01
- 9.44417724e-01 9.51463269e-01 9.58054229e-01 9.64183967e-01
- 9.69846310e-01 9.75035559e-01 9.79746487e-01 9.83974351e-01
- 9.87714893e-01 9.90964349e-01 9.93719444e-01 9.95977406e-01
- 9.97735961e-01 9.98993338e-01 9.99748271e-01 1.00000000e+00
- 9.99748271e-01 9.98993338e-01 9.97735961e-01 9.95977406e-01
- 9.93719444e-01 9.90964349e-01 9.87714893e-01 9.83974351e-01
- 9.79746487e-01 9.75035559e-01 9.69846310e-01 9.64183967e-01
- 9.58054229e-01 9.51463269e-01 9.44417724e-01 9.36924689e-01
- 9.28991707e-01 9.20626766e-01 9.11838291e-01 9.02635129e-01
- 8.93026547e-01 8.83022222e-01 8.72632225e-01 8.61867019e-01
- 8.50737444e-01 8.39254706e-01 8.27430367e-01 8.15276334e-01
- 8.02804844e-01 7.90028455e-01 7.76960032e-01 7.63612734e-01
- 7.50000000e-01 7.36135537e-01 7.22033306e-01 7.07707507e-01
- 6.93172563e-01 6.78443111e-01 6.63533982e-01 6.48460188e-01
- 6.33236907e-01 6.17879468e-01 6.02403334e-01 5.86824089e-01
- 5.71157419e-01 5.55419100e-01 5.39624978e-01 5.23790958e-01
- 5.07932982e-01 4.92067018e-01 4.76209042e-01 4.60375022e-01
- 4.44580900e-01 4.28842581e-01 4.13175911e-01 3.97596666e-01
- 3.82120532e-01 3.66763093e-01 3.51539812e-01 3.36466018e-01
- 3.21556889e-01 3.06827437e-01 2.92292493e-01 2.77966694e-01
- 2.63864463e-01 2.50000000e-01 2.36387266e-01 2.23039968e-01
- 2.09971545e-01 1.97195156e-01 1.84723666e-01 1.72569633e-01
- 1.60745294e-01 1.49262556e-01 1.38132981e-01 1.27367775e-01
- 1.16977778e-01 1.06973453e-01 9.73648712e-02 8.81617093e-02
- 7.93732336e-02 7.10082934e-02 6.30753115e-02 5.55822757e-02
- 4.85367309e-02 4.19457713e-02 3.58160335e-02 3.01536896e-02
- 2.49644411e-02 2.02535132e-02 1.60256493e-02 1.22851066e-02
- 9.03565137e-03 6.28055566e-03 4.02259358e-03 2.26403871e-03
- 1.00666176e-03 2.51728808e-04] [ 0.          0.01462169  0.03093556  0.04686826  0.06258357  0.07861745
-  0.09405252  0.10967754  0.12517467  0.14030205  0.15556546  0.17061296
-  0.18530152  0.19996301  0.21449599  0.22874252  0.24262855  0.25634017
-  0.26983858  0.28309181  0.29606886  0.30870461  0.32101182  0.33297108
-  0.34460585  0.35590175  0.36684564  0.37742782  0.38764095  0.39746876
-  0.40687689  0.4158685   0.4244587   0.43265564  0.44035434  0.44758788
-  0.45439201  0.4607208   0.4666015   0.4720264   0.47698672  0.48148686
-  0.48552288  0.48903208  0.49198712  0.4944485   0.49644181  0.49797497
-  0.49903068  0.49956586  0.49951122  0.49896566  0.49795529  0.496483
-  0.49454021  0.49197826  0.488932    0.48544118  0.4815031   0.47706314
-  0.47204391  0.46659337  0.46072373  0.45444908  0.44755198  0.44023838
-  0.4325517   0.42450843  0.41583676  0.40678988  0.39741734  0.38773107
-  0.37741607  0.36678968  0.35588415  0.34465455  0.33287266  0.32086115
-  0.30864941  0.29605157  0.28300335  0.26980349  0.25650823  0.24258332
-  0.22849786  0.21437412  0.20008925  0.18538938  0.17029503  0.15488099
-  0.13980699  0.12467987  0.10951799  0.09361538  0.07689355  0.06143269
-  0.04835402  0.03087902  0.0122301   0.00369928 -0.01428549 -0.03074717
- -0.04732177 -0.06375087 -0.07907204 -0.09463611 -0.11041152 -0.12590768
- -0.14098449 -0.15620072 -0.17142561 -0.18610004 -0.20064589 -0.21515702
- -0.2295169  -0.24337121 -0.2570841  -0.27063306 -0.28390427 -0.29672722
- -0.30934207 -0.32172203 -0.33375775 -0.34532288 -0.35662667 -0.36763219
- -0.37826137 -0.38840314 -0.39819124 -0.40761009 -0.41665301 -0.4252099
- -0.43334852 -0.4410498  -0.44831686 -0.45518375 -0.46154504 -0.46740234
- -0.47283049 -0.47778693 -0.48225506 -0.48627201 -0.48977107 -0.49274869
- -0.49525282 -0.49725252 -0.49875597 -0.49976917 -0.50027517 -0.50029574
- -0.49981064 -0.4988094  -0.49728292 -0.49526513 -0.49277202 -0.48977413
- -0.48627141 -0.482278   -0.47780307 -0.47284984 -0.46742386 -0.46153196
- -0.45518172 -0.4483774  -0.44107921 -0.43334623 -0.4251908  -0.41662661
- -0.40766583 -0.39826616 -0.38843777 -0.37820676 -0.36765976 -0.35672833
- -0.34538119 -0.33373416 -0.32178271 -0.30940667 -0.29682167 -0.28386015
- -0.27065183 -0.25715717 -0.24348009 -0.22943138 -0.21522871 -0.20084706
- -0.18616117 -0.17135669 -0.15630674 -0.14115807 -0.12588878 -0.11053819
- -0.09491395 -0.0792819  -0.06370407 -0.04766179 -0.03168923 -0.01532304] [3440000000.0, 41800000000.0, 27700000000.0, 13600000000.0, 200000000000.0, 205000000000.0, 118000000000.0, 98460000.0, 114500000000.0, 1000000.0, 4560000.0] [3440000000.0, 14000000000.0, 13650000000.0, 13300000000.0, 200000000000.0, 205000000000.0, 118000000000.0, 98460000.0, 8390000000.0, 1000000.0, 4560000.0] [1323000000.0, 2630000000.0, 7200000000.0, 11800000000.0, 79300000000.0, 80000000000.0, 47600000000.0, 98460000.0, 5990000000.0, 384615.3846153846, 1520000.0] [0.3, 0.28, 0.39, 0.49, 0.3, 0.3, 0.3, 0.3, 0.27, 0.3, 0.49] [1235.0, 1920.0, 1850.0, 1780.0, 7800.0, 7850.0, 7200.0, 200.0, 1220.0, 1150.0, 1100.0] [0.         0.62506011 0.81300099 1.        ] [4. 5. 4.] [1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1.] [5.00e-05 2.82e-03 5.17e-02 2.82e-03 5.00e-05 2.82e-03 5.17e-02 4.70e-04
- 2.82e-03 5.00e-05 2.82e-03 5.17e-02 2.82e-03] [0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0.] [1 3 3 3 1 3 3 9 3 1 3 3 3] [0.         0.18705037 0.37497357 1.        ] [4. 5. 4.] [1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1.] [5.00e-05 2.82e-03 5.17e-02 2.82e-03 5.00e-05 2.82e-03 5.17e-02 4.70e-04
- 2.82e-03 5.00e-05 2.82e-03 5.17e-02 2.82e-03] [0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0.] [1 3 3 3 1 3 3 9 3 1 3 3 3] 2 [0.64505453 0.79653854] [3. 3.] [1. 1. 1. 1. 1. 1.] [0.002 0.04  0.002 0.002 0.04  0.002] [0. 0. 0. 0. 0. 0.] [4 8 4 4 8 4]
-
-1 3.5829065688610924 0.0 0.0 0.4915122797251875 [0.00000000e+00 2.51728808e-04 1.00666176e-03 2.26403871e-03
- 4.02259358e-03 6.28055566e-03 9.03565137e-03 1.22851066e-02
- 1.60256493e-02 2.02535132e-02 2.49644411e-02 3.01536896e-02
- 3.58160335e-02 4.19457713e-02 4.85367309e-02 5.55822757e-02
- 6.30753115e-02 7.10082934e-02 7.93732336e-02 8.81617093e-02
- 9.73648712e-02 1.06973453e-01 1.16977778e-01 1.27367775e-01
- 1.38132981e-01 1.49262556e-01 1.60745294e-01 1.72569633e-01
- 1.84723666e-01 1.97195156e-01 2.09971545e-01 2.23039968e-01
- 2.36387266e-01 2.50000000e-01 2.63864463e-01 2.77966694e-01
- 2.92292493e-01 3.06827437e-01 3.21556889e-01 3.36466018e-01
- 3.51539812e-01 3.66763093e-01 3.82120532e-01 3.97596666e-01
- 4.13175911e-01 4.28842581e-01 4.44580900e-01 4.60375022e-01
- 4.76209042e-01 4.92067018e-01 5.07932982e-01 5.23790958e-01
- 5.39624978e-01 5.55419100e-01 5.71157419e-01 5.86824089e-01
- 6.02403334e-01 6.17879468e-01 6.33236907e-01 6.48460188e-01
- 6.63533982e-01 6.78443111e-01 6.93172563e-01 7.07707507e-01
- 7.22033306e-01 7.36135537e-01 7.50000000e-01 7.63612734e-01
- 7.76960032e-01 7.90028455e-01 8.02804844e-01 8.15276334e-01
- 8.27430367e-01 8.39254706e-01 8.50737444e-01 8.61867019e-01
- 8.72632225e-01 8.83022222e-01 8.93026547e-01 9.02635129e-01
- 9.11838291e-01 9.20626766e-01 9.28991707e-01 9.36924689e-01
- 9.44417724e-01 9.51463269e-01 9.58054229e-01 9.64183967e-01
- 9.69846310e-01 9.75035559e-01 9.79746487e-01 9.83974351e-01
- 9.87714893e-01 9.90964349e-01 9.93719444e-01 9.95977406e-01
- 9.97735961e-01 9.98993338e-01 9.99748271e-01 1.00000000e+00
- 9.99748271e-01 9.98993338e-01 9.97735961e-01 9.95977406e-01
- 9.93719444e-01 9.90964349e-01 9.87714893e-01 9.83974351e-01
- 9.79746487e-01 9.75035559e-01 9.69846310e-01 9.64183967e-01
- 9.58054229e-01 9.51463269e-01 9.44417724e-01 9.36924689e-01
- 9.28991707e-01 9.20626766e-01 9.11838291e-01 9.02635129e-01
- 8.93026547e-01 8.83022222e-01 8.72632225e-01 8.61867019e-01
- 8.50737444e-01 8.39254706e-01 8.27430367e-01 8.15276334e-01
- 8.02804844e-01 7.90028455e-01 7.76960032e-01 7.63612734e-01
- 7.50000000e-01 7.36135537e-01 7.22033306e-01 7.07707507e-01
- 6.93172563e-01 6.78443111e-01 6.63533982e-01 6.48460188e-01
- 6.33236907e-01 6.17879468e-01 6.02403334e-01 5.86824089e-01
- 5.71157419e-01 5.55419100e-01 5.39624978e-01 5.23790958e-01
- 5.07932982e-01 4.92067018e-01 4.76209042e-01 4.60375022e-01
- 4.44580900e-01 4.28842581e-01 4.13175911e-01 3.97596666e-01
- 3.82120532e-01 3.66763093e-01 3.51539812e-01 3.36466018e-01
- 3.21556889e-01 3.06827437e-01 2.92292493e-01 2.77966694e-01
- 2.63864463e-01 2.50000000e-01 2.36387266e-01 2.23039968e-01
- 2.09971545e-01 1.97195156e-01 1.84723666e-01 1.72569633e-01
- 1.60745294e-01 1.49262556e-01 1.38132981e-01 1.27367775e-01
- 1.16977778e-01 1.06973453e-01 9.73648712e-02 8.81617093e-02
- 7.93732336e-02 7.10082934e-02 6.30753115e-02 5.55822757e-02
- 4.85367309e-02 4.19457713e-02 3.58160335e-02 3.01536896e-02
- 2.49644411e-02 2.02535132e-02 1.60256493e-02 1.22851066e-02
- 9.03565137e-03 6.28055566e-03 4.02259358e-03 2.26403871e-03
- 1.00666176e-03 2.51728808e-04] [ 0.          0.01460331  0.0309479   0.04688173  0.06262183  0.07852066
-  0.09366058  0.10892945  0.12431392  0.13951841  0.15487148  0.17011078
-  0.18489809  0.1996642   0.2143079   0.22871394  0.24271145  0.25653524
-  0.27010459  0.28337814  0.2963693   0.30880376  0.32087878  0.33259465
-  0.34397545  0.35501369  0.36570174  0.37603022  0.38599634  0.39558804
-  0.40477526  0.41355431  0.42194275  0.42991362  0.43741186  0.44447547
-  0.45111737  0.45730268  0.46304193  0.46833484  0.47317083  0.47755375
-  0.48149045  0.48482756  0.48770166  0.49009639  0.49202843  0.49351291
-  0.49454481  0.49496787  0.49489874  0.49436     0.49335904  0.49189959
-  0.48987108  0.48727799  0.48423013  0.48073995  0.47681362  0.47230071
-  0.46730854  0.46189391  0.45606608  0.44978422  0.44293089  0.43569121
-  0.42808615  0.42010608  0.41150619  0.40258074  0.39334522  0.38375571
-  0.37360136  0.36316835  0.35247728  0.34139921  0.32988229  0.3181616
-  0.30627292  0.29393234  0.28129741  0.26854951  0.2556691   0.24225974
-  0.22883033  0.21546372  0.20171269  0.18797646  0.1743387   0.15989881
-  0.14594316  0.1327296   0.11972514  0.1047219   0.08870146  0.07286023
-  0.05781044  0.04290815  0.02863671  0.00664756 -0.00129742 -0.01942787
- -0.03638897 -0.05115015 -0.06698504 -0.08314182 -0.09872811 -0.11399554
- -0.12949067 -0.14506465 -0.16008784 -0.17497738 -0.18986304 -0.20467094
- -0.21893163 -0.23308644 -0.24710315 -0.26090589 -0.27421601 -0.28730694
- -0.30017523 -0.31278    -0.32492593 -0.33671781 -0.34822897 -0.35942891
- -0.37020166 -0.38053041 -0.39048859 -0.40003469 -0.40916003 -0.41780257
- -0.42604362 -0.43385624 -0.44123963 -0.4482293  -0.45471438 -0.46071088
- -0.46628389 -0.47139264 -0.47601911 -0.48020416 -0.48387633 -0.48705124
- -0.48976445 -0.49198897 -0.49373217 -0.49499841 -0.49576984 -0.49607691
- -0.49589231 -0.49519644 -0.49399141 -0.49230472 -0.49014008 -0.48747007
- -0.48428845 -0.48060382 -0.47642307 -0.47174478 -0.46658122 -0.46093461
- -0.45481164 -0.44820564 -0.44110377 -0.43355038 -0.42555989 -0.41714476
- -0.40822907 -0.39881035 -0.38894368 -0.37865105 -0.36803624 -0.35703868
- -0.34561479 -0.33391915 -0.32192031 -0.30948905 -0.29687954 -0.28387762
- -0.27073409 -0.25735359 -0.2437707  -0.22970784 -0.21548402 -0.2010934
- -0.1863887  -0.17156798 -0.15649765 -0.1413169  -0.12603923 -0.11067484
- -0.09498058 -0.07935878 -0.06379848 -0.04773762 -0.03174707 -0.01536291] [3440000000.0, 41800000000.0, 27700000000.0, 13600000000.0, 200000000000.0, 205000000000.0, 118000000000.0, 98460000.0, 114500000000.0, 1000000.0, 4560000.0] [3440000000.0, 14000000000.0, 13650000000.0, 13300000000.0, 200000000000.0, 205000000000.0, 118000000000.0, 98460000.0, 8390000000.0, 1000000.0, 4560000.0] [1323000000.0, 2630000000.0, 7200000000.0, 11800000000.0, 79300000000.0, 80000000000.0, 47600000000.0, 98460000.0, 5990000000.0, 384615.3846153846, 1520000.0] [0.3, 0.28, 0.39, 0.49, 0.3, 0.3, 0.3, 0.3, 0.27, 0.3, 0.49] [1235.0, 1920.0, 1850.0, 1780.0, 7800.0, 7850.0, 7200.0, 200.0, 1220.0, 1150.0, 1100.0] [0.         0.61627184 0.80384044 0.99952675 1.        ] [4. 5. 5. 5.] [1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1.] [5.00000000e-05 2.82000000e-03 2.59593191e-02 2.82000000e-03
- 5.00000000e-05 2.82000000e-03 2.59593191e-02 2.63777678e-03
- 2.82000000e-03 5.00000000e-05 2.82000000e-03 2.59593191e-02
- 2.00000000e-02 2.82000000e-03 5.00000000e-05 2.82000000e-03
- 2.59593191e-02 1.76336508e-03 2.82000000e-03] [0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0.] [1 3 3 3 1 3 3 9 3 1 3 3 8 3 1 3 3 2 3] [0.         0.18267241 0.36875322 0.94281614 1.        ] [4. 5. 5. 5.] [1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1.] [5.00000000e-05 2.82000000e-03 2.59593191e-02 2.82000000e-03
- 5.00000000e-05 2.82000000e-03 2.59593191e-02 2.63777678e-03
- 2.82000000e-03 5.00000000e-05 2.82000000e-03 2.59593191e-02
- 2.00000000e-02 2.82000000e-03 5.00000000e-05 2.82000000e-03
- 2.59593191e-02 1.76336508e-03 2.82000000e-03] [0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0.] [1 3 3 3 1 3 3 9 3 1 3 3 8 3 1 3 3 2 3] 2 [0.63605513 0.78722441] [3. 3.] [1. 1. 1. 1. 1. 1.] [0.002 0.04  0.002 0.002 0.04  0.002] [0. 0. 0. 0. 0. 0.] [4 8 4 4 8 4]
-
-
-2 3.8701677888975055 0.0 0.0 0.4676395111012256 [0.00000000e+00 2.51728808e-04 1.00666176e-03 2.26403871e-03
- 4.02259358e-03 6.28055566e-03 9.03565137e-03 1.22851066e-02
- 1.60256493e-02 2.02535132e-02 2.49644411e-02 3.01536896e-02
- 3.58160335e-02 4.19457713e-02 4.85367309e-02 5.55822757e-02
- 6.30753115e-02 7.10082934e-02 7.93732336e-02 8.81617093e-02
- 9.73648712e-02 1.06973453e-01 1.16977778e-01 1.27367775e-01
- 1.38132981e-01 1.49262556e-01 1.60745294e-01 1.72569633e-01
- 1.84723666e-01 1.97195156e-01 2.09971545e-01 2.23039968e-01
- 2.36387266e-01 2.50000000e-01 2.63864463e-01 2.77966694e-01
- 2.92292493e-01 3.06827437e-01 3.21556889e-01 3.36466018e-01
- 3.51539812e-01 3.66763093e-01 3.82120532e-01 3.97596666e-01
- 4.13175911e-01 4.28842581e-01 4.44580900e-01 4.60375022e-01
- 4.76209042e-01 4.92067018e-01 5.07932982e-01 5.23790958e-01
- 5.39624978e-01 5.55419100e-01 5.71157419e-01 5.86824089e-01
- 6.02403334e-01 6.17879468e-01 6.33236907e-01 6.48460188e-01
- 6.63533982e-01 6.78443111e-01 6.93172563e-01 7.07707507e-01
- 7.22033306e-01 7.36135537e-01 7.50000000e-01 7.63612734e-01
- 7.76960032e-01 7.90028455e-01 8.02804844e-01 8.15276334e-01
- 8.27430367e-01 8.39254706e-01 8.50737444e-01 8.61867019e-01
- 8.72632225e-01 8.83022222e-01 8.93026547e-01 9.02635129e-01
- 9.11838291e-01 9.20626766e-01 9.28991707e-01 9.36924689e-01
- 9.44417724e-01 9.51463269e-01 9.58054229e-01 9.64183967e-01
- 9.69846310e-01 9.75035559e-01 9.79746487e-01 9.83974351e-01
- 9.87714893e-01 9.90964349e-01 9.93719444e-01 9.95977406e-01
- 9.97735961e-01 9.98993338e-01 9.99748271e-01 1.00000000e+00
- 9.99748271e-01 9.98993338e-01 9.97735961e-01 9.95977406e-01
- 9.93719444e-01 9.90964349e-01 9.87714893e-01 9.83974351e-01
- 9.79746487e-01 9.75035559e-01 9.69846310e-01 9.64183967e-01
- 9.58054229e-01 9.51463269e-01 9.44417724e-01 9.36924689e-01
- 9.28991707e-01 9.20626766e-01 9.11838291e-01 9.02635129e-01
- 8.93026547e-01 8.83022222e-01 8.72632225e-01 8.61867019e-01
- 8.50737444e-01 8.39254706e-01 8.27430367e-01 8.15276334e-01
- 8.02804844e-01 7.90028455e-01 7.76960032e-01 7.63612734e-01
- 7.50000000e-01 7.36135537e-01 7.22033306e-01 7.07707507e-01
- 6.93172563e-01 6.78443111e-01 6.63533982e-01 6.48460188e-01
- 6.33236907e-01 6.17879468e-01 6.02403334e-01 5.86824089e-01
- 5.71157419e-01 5.55419100e-01 5.39624978e-01 5.23790958e-01
- 5.07932982e-01 4.92067018e-01 4.76209042e-01 4.60375022e-01
- 4.44580900e-01 4.28842581e-01 4.13175911e-01 3.97596666e-01
- 3.82120532e-01 3.66763093e-01 3.51539812e-01 3.36466018e-01
- 3.21556889e-01 3.06827437e-01 2.92292493e-01 2.77966694e-01
- 2.63864463e-01 2.50000000e-01 2.36387266e-01 2.23039968e-01
- 2.09971545e-01 1.97195156e-01 1.84723666e-01 1.72569633e-01
- 1.60745294e-01 1.49262556e-01 1.38132981e-01 1.27367775e-01
- 1.16977778e-01 1.06973453e-01 9.73648712e-02 8.81617093e-02
- 7.93732336e-02 7.10082934e-02 6.30753115e-02 5.55822757e-02
- 4.85367309e-02 4.19457713e-02 3.58160335e-02 3.01536896e-02
- 2.49644411e-02 2.02535132e-02 1.60256493e-02 1.22851066e-02
- 9.03565137e-03 6.28055566e-03 4.02259358e-03 2.26403871e-03
- 1.00666176e-03 2.51728808e-04]  [ 0.          0.01120969  0.02701929  0.04218065  0.05764482  0.07246875
-  0.08564168  0.09834576  0.11148654  0.12550799  0.1411993   0.15690758
-  0.17224704  0.18747775  0.20256409  0.21735406  0.23191916  0.24626673
-  0.26013402  0.27301756  0.28552724  0.29676874  0.30665993  0.31617414
-  0.32532549  0.33411744  0.34254552  0.35061658  0.35834619  0.36574538
-  0.37283081  0.37959905  0.38603872  0.39215614  0.3979167   0.40331907
-  0.40835325  0.41299069  0.41725138  0.42110714  0.4245553   0.42760207
-  0.43025379  0.43250972  0.4343545   0.43575909  0.43673921  0.43734126
-  0.43757783  0.43743239  0.43684496  0.43580242  0.43436941  0.43250965
-  0.43012006  0.42705414  0.42353895  0.41954338  0.41513615  0.41034702
-  0.40519817  0.39969844  0.39381566  0.38751961  0.38088845  0.37394349
-  0.36671061  0.35921499  0.35130023  0.34313941  0.33476673  0.32619607
-  0.31740772  0.30833978  0.2991073   0.28976164  0.2803476   0.27070275
-  0.26107339  0.2514803   0.2418014   0.23212832  0.2226081   0.21316986
-  0.20381882  0.19480624  0.18597175  0.17745396  0.16934226  0.1613715
-  0.15393284  0.14708226  0.14095694  0.1355312   0.13060215  0.1265625
-  0.12341632  0.12116678  0.11840959  0.11734537  0.11527624  0.10907095
-  0.09960581  0.08928819  0.07724998  0.0645792   0.05144824  0.03755227
-  0.02318632  0.00849216 -0.0065424  -0.0220048  -0.03779039 -0.05382507
- -0.07011082 -0.08657413 -0.10310044 -0.11962767 -0.13608252 -0.15239191
- -0.16849302 -0.1842925  -0.1998131  -0.21501406 -0.2297506  -0.24409529
- -0.25800018 -0.27134534 -0.28420271 -0.29664822 -0.30844947 -0.31910191
- -0.32918323 -0.33892513 -0.34829371 -0.35715935 -0.36554042 -0.37355249
- -0.38114172 -0.38830568 -0.39503232 -0.40129157 -0.40716817 -0.41263096
- -0.41773261 -0.42246065 -0.4268233  -0.43083607 -0.43445839 -0.43772633
- -0.44068417 -0.44323915 -0.44535791 -0.44707321 -0.44833175 -0.44904395
- -0.44919423 -0.44871379 -0.44758109 -0.44577392 -0.44331975 -0.44023795
- -0.43650598 -0.43217939 -0.42724579 -0.42164557 -0.41547069 -0.40873092
- -0.40078507 -0.39181058 -0.38238933 -0.37209767 -0.36140497 -0.3503803
- -0.33895659 -0.32730461 -0.31537688 -0.30312244 -0.29074482 -0.27803337
- -0.26580144 -0.25383182 -0.24140272 -0.2278797  -0.21407735 -0.20000771
- -0.18568482 -0.1712052  -0.15649897 -0.14158192 -0.12667381 -0.11180655
- -0.09598118 -0.08049236 -0.06572959 -0.05022819 -0.03484757 -0.01948332] [3440000000.0, 41800000000.0, 27700000000.0, 13600000000.0, 200000000000.0, 205000000000.0, 118000000000.0, 98460000.0, 114500000000.0, 1000000.0, 4560000.0] [3440000000.0, 14000000000.0, 13650000000.0, 13300000000.0, 200000000000.0, 205000000000.0, 118000000000.0, 98460000.0, 8390000000.0, 1000000.0, 4560000.0] [1323000000.0, 2630000000.0, 7200000000.0, 11800000000.0, 79300000000.0, 80000000000.0, 47600000000.0, 98460000.0, 5990000000.0, 384615.3846153846, 1520000.0] [0.3, 0.28, 0.39, 0.49, 0.3, 0.3, 0.3, 0.3, 0.27, 0.3, 0.49] [1235.0, 1920.0, 1850.0, 1780.0, 7800.0, 7850.0, 7200.0, 200.0, 1220.0, 1150.0, 1100.0] [0.         0.57161483 0.74719882 0.9731418  1.        ] [4. 5. 5. 5.] [1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1.] [5.00000000e-05 2.82000000e-03 1.24339293e-02 2.82000000e-03
- 5.00000000e-05 2.82000000e-03 1.24339293e-02 2.30062882e-02
- 2.82000000e-03 5.00000000e-05 2.82000000e-03 1.24339293e-02
- 2.00000000e-02 2.82000000e-03 5.00000000e-05 2.82000000e-03
- 1.24339293e-02 6.03688768e-03 2.82000000e-03] [0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0.] [1 3 3 3 1 3 3 9 3 1 3 3 8 3 1 3 3 2 3] [0.         0.18292683 0.36533222 0.93287885 1.        ] [4. 5. 5. 5.] [1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1.] [5.00000000e-05 2.82000000e-03 1.24339293e-02 2.82000000e-03
- 5.00000000e-05 2.82000000e-03 1.24339293e-02 2.30062882e-02
- 2.82000000e-03 5.00000000e-05 2.82000000e-03 1.24339293e-02
- 2.00000000e-02 2.82000000e-03 5.00000000e-05 2.82000000e-03
- 1.24339293e-02 6.03688768e-03 2.82000000e-03] [0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0.] [1 3 3 3 1 3 3 9 3 1 3 3 8 3 1 3 3 2 3] 2 [0.59002035 0.73118979] [3. 3.] [1. 1. 1. 1. 1. 1.] [0.002 0.04  0.002 0.002 0.04  0.002] [0. 0. 0. 0. 0. 0.] [4 8 4 4 8 4]
-
-
-3 4.113082662273744 0.0 0.0 0.4437667424772636 [0.00000000e+00 2.51728808e-04 1.00666176e-03 2.26403871e-03
- 4.02259358e-03 6.28055566e-03 9.03565137e-03 1.22851066e-02
- 1.60256493e-02 2.02535132e-02 2.49644411e-02 3.01536896e-02
- 3.58160335e-02 4.19457713e-02 4.85367309e-02 5.55822757e-02
- 6.30753115e-02 7.10082934e-02 7.93732336e-02 8.81617093e-02
- 9.73648712e-02 1.06973453e-01 1.16977778e-01 1.27367775e-01
- 1.38132981e-01 1.49262556e-01 1.60745294e-01 1.72569633e-01
- 1.84723666e-01 1.97195156e-01 2.09971545e-01 2.23039968e-01
- 2.36387266e-01 2.50000000e-01 2.63864463e-01 2.77966694e-01
- 2.92292493e-01 3.06827437e-01 3.21556889e-01 3.36466018e-01
- 3.51539812e-01 3.66763093e-01 3.82120532e-01 3.97596666e-01
- 4.13175911e-01 4.28842581e-01 4.44580900e-01 4.60375022e-01
- 4.76209042e-01 4.92067018e-01 5.07932982e-01 5.23790958e-01
- 5.39624978e-01 5.55419100e-01 5.71157419e-01 5.86824089e-01
- 6.02403334e-01 6.17879468e-01 6.33236907e-01 6.48460188e-01
- 6.63533982e-01 6.78443111e-01 6.93172563e-01 7.07707507e-01
- 7.22033306e-01 7.36135537e-01 7.50000000e-01 7.63612734e-01
- 7.76960032e-01 7.90028455e-01 8.02804844e-01 8.15276334e-01
- 8.27430367e-01 8.39254706e-01 8.50737444e-01 8.61867019e-01
- 8.72632225e-01 8.83022222e-01 8.93026547e-01 9.02635129e-01
- 9.11838291e-01 9.20626766e-01 9.28991707e-01 9.36924689e-01
- 9.44417724e-01 9.51463269e-01 9.58054229e-01 9.64183967e-01
- 9.69846310e-01 9.75035559e-01 9.79746487e-01 9.83974351e-01
- 9.87714893e-01 9.90964349e-01 9.93719444e-01 9.95977406e-01
- 9.97735961e-01 9.98993338e-01 9.99748271e-01 1.00000000e+00
- 9.99748271e-01 9.98993338e-01 9.97735961e-01 9.95977406e-01
- 9.93719444e-01 9.90964349e-01 9.87714893e-01 9.83974351e-01
- 9.79746487e-01 9.75035559e-01 9.69846310e-01 9.64183967e-01
- 9.58054229e-01 9.51463269e-01 9.44417724e-01 9.36924689e-01
- 9.28991707e-01 9.20626766e-01 9.11838291e-01 9.02635129e-01
- 8.93026547e-01 8.83022222e-01 8.72632225e-01 8.61867019e-01
- 8.50737444e-01 8.39254706e-01 8.27430367e-01 8.15276334e-01
- 8.02804844e-01 7.90028455e-01 7.76960032e-01 7.63612734e-01
- 7.50000000e-01 7.36135537e-01 7.22033306e-01 7.07707507e-01
- 6.93172563e-01 6.78443111e-01 6.63533982e-01 6.48460188e-01
- 6.33236907e-01 6.17879468e-01 6.02403334e-01 5.86824089e-01
- 5.71157419e-01 5.55419100e-01 5.39624978e-01 5.23790958e-01
- 5.07932982e-01 4.92067018e-01 4.76209042e-01 4.60375022e-01
- 4.44580900e-01 4.28842581e-01 4.13175911e-01 3.97596666e-01
- 3.82120532e-01 3.66763093e-01 3.51539812e-01 3.36466018e-01
- 3.21556889e-01 3.06827437e-01 2.92292493e-01 2.77966694e-01
- 2.63864463e-01 2.50000000e-01 2.36387266e-01 2.23039968e-01
- 2.09971545e-01 1.97195156e-01 1.84723666e-01 1.72569633e-01
- 1.60745294e-01 1.49262556e-01 1.38132981e-01 1.27367775e-01
- 1.16977778e-01 1.06973453e-01 9.73648712e-02 8.81617093e-02
- 7.93732336e-02 7.10082934e-02 6.30753115e-02 5.55822757e-02
- 4.85367309e-02 4.19457713e-02 3.58160335e-02 3.01536896e-02
- 2.49644411e-02 2.02535132e-02 1.60256493e-02 1.22851066e-02
- 9.03565137e-03 6.28055566e-03 4.02259358e-03 2.26403871e-03
- 1.00666176e-03 2.51728808e-04]  [ 0.          0.01121936  0.02514038  0.03854025  0.05184036  0.06496535
-  0.0780105   0.08957465  0.10039942  0.11139644  0.1234451   0.1369198
-  0.15044289  0.16382788  0.17691041  0.18987833  0.20258004  0.21499631
-  0.22718769  0.23892747  0.24949197  0.25966298  0.26871493  0.27651932
-  0.28392681  0.29096256  0.29762221  0.30391418  0.30983635  0.31542347
-  0.32068054  0.32566486  0.33035227  0.33475793  0.33884298  0.34261573
-  0.34607024  0.34920105  0.35202419  0.3545077   0.35660897  0.3583414
-  0.35971324  0.36072475  0.36138421  0.36170775  0.3616423   0.36125352
-  0.36054714  0.35948217  0.35807254  0.35634649  0.35427139  0.35177472
-  0.34857337  0.34506383  0.34122228  0.33705907  0.3326567   0.32803248
-  0.32317117  0.31809356  0.3128334   0.30738996  0.30178512  0.29602361
-  0.29011509  0.28411517  0.27805171  0.27194112  0.26579902  0.25963489
-  0.25339835  0.2472072   0.24108095  0.23502735  0.22909013  0.22324402
-  0.21753015  0.21197461  0.20656834  0.20136496  0.1963648   0.19160083
-  0.18708072  0.18275153  0.17867132  0.17480316  0.171203    0.16790332
-  0.16490741  0.1622187   0.15968204  0.15736863  0.15540717  0.15379963
-  0.15254765  0.15009149  0.14689989  0.14583567  0.14562749  0.14500318
-  0.14396335  0.14250906  0.14064178  0.13836337  0.13592321  0.13319317
-  0.13009552  0.12649018  0.12251877  0.11787687  0.11276142  0.10690919
-  0.10045748  0.09340274  0.0856032   0.07697958  0.06756896  0.05729537
-  0.04611424  0.03402371  0.02097113  0.00706204 -0.00759932 -0.02291182
- -0.03869128 -0.05477656 -0.0710427  -0.08739384 -0.10348447 -0.11936972
- -0.1348852  -0.14996108 -0.16466431 -0.17871382 -0.19110948 -0.20309572
- -0.21463229 -0.22558228 -0.23602788 -0.24605102 -0.25553062 -0.26452767
- -0.27297968 -0.28103777 -0.28875119 -0.29614921 -0.30318341 -0.30992051
- -0.3163293  -0.32233225 -0.32811701 -0.33348773 -0.33835887 -0.34279527
- -0.34665362 -0.34983373 -0.3522932  -0.35394853 -0.35479787 -0.35474149
- -0.35397192 -0.35243934 -0.3502157  -0.34733827 -0.34379272 -0.33957638
- -0.33471871 -0.32873158 -0.32163836 -0.31407208 -0.30545927 -0.29639508
- -0.28697433 -0.27721027 -0.26722647 -0.25690341 -0.24641366 -0.23570344
- -0.22484702 -0.21462195 -0.20476813 -0.19422563 -0.18259411 -0.17070717
- -0.15865639 -0.14626428 -0.13366701 -0.12075235 -0.10766342 -0.09490202
- -0.08195323 -0.06760886 -0.05387712 -0.04134647 -0.02824176 -0.01519076] [3440000000.0, 41800000000.0, 27700000000.0, 13600000000.0, 200000000000.0, 205000000000.0, 118000000000.0, 98460000.0, 114500000000.0, 1000000.0, 4560000.0] [3440000000.0, 14000000000.0, 13650000000.0, 13300000000.0, 200000000000.0, 205000000000.0, 118000000000.0, 98460000.0, 8390000000.0, 1000000.0, 4560000.0] [1323000000.0, 2630000000.0, 7200000000.0, 11800000000.0, 79300000000.0, 80000000000.0, 47600000000.0, 98460000.0, 5990000000.0, 384615.3846153846, 1520000.0] [0.3, 0.28, 0.39, 0.49, 0.3, 0.3, 0.3, 0.3, 0.27, 0.3, 0.49] [1235.0, 1920.0, 1850.0, 1780.0, 7800.0, 7850.0, 7200.0, 200.0, 1220.0, 1150.0, 1100.0] [0.         0.50678567 0.66728724 0.91646694 1.        ] [4. 5. 5. 5.] [1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1.] [5.00000000e-05 2.82000000e-03 4.27131105e-03 2.82000000e-03
- 5.00000000e-05 2.82000000e-03 4.27131105e-03 4.14211021e-02
- 2.82000000e-03 5.00000000e-05 2.82000000e-03 4.27131105e-03
- 6.99488261e-02 2.82000000e-03 5.00000000e-05 2.82000000e-03
- 4.27131105e-03 7.05000000e-03 2.82000000e-03] [0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0.] [1 3 3 3 1 3 3 9 3 1 3 3 8 3 1 3 3 2 3] [0.         0.19761921 0.37702993 0.90430839 1.        ] [4. 5. 5. 5.] [1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1.] [5.00000000e-05 2.82000000e-03 4.27131105e-03 2.82000000e-03
- 5.00000000e-05 2.82000000e-03 4.27131105e-03 4.14211021e-02
- 2.82000000e-03 5.00000000e-05 2.82000000e-03 4.27131105e-03
- 6.99488261e-02 2.82000000e-03 5.00000000e-05 2.82000000e-03
- 4.27131105e-03 7.05000000e-03 2.82000000e-03] [0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0.] [1 3 3 3 1 3 3 9 3 1 3 3 8 3 1 3 3 2 3] 2 [0.52371793 0.65273678] [3. 3.] [1. 1. 1. 1. 1. 1.] [0.002 0.04  0.002 0.002 0.04  0.002] [0. 0. 0. 0. 0. 0.] [4 8 4 4 8 4]
-
-
-4 4.386184316700438 0.0 0.0 0.4198939738533017 [0.00000000e+00 2.51728808e-04 1.00666176e-03 2.26403871e-03
- 4.02259358e-03 6.28055566e-03 9.03565137e-03 1.22851066e-02
- 1.60256493e-02 2.02535132e-02 2.49644411e-02 3.01536896e-02
- 3.58160335e-02 4.19457713e-02 4.85367309e-02 5.55822757e-02
- 6.30753115e-02 7.10082934e-02 7.93732336e-02 8.81617093e-02
- 9.73648712e-02 1.06973453e-01 1.16977778e-01 1.27367775e-01
- 1.38132981e-01 1.49262556e-01 1.60745294e-01 1.72569633e-01
- 1.84723666e-01 1.97195156e-01 2.09971545e-01 2.23039968e-01
- 2.36387266e-01 2.50000000e-01 2.63864463e-01 2.77966694e-01
- 2.92292493e-01 3.06827437e-01 3.21556889e-01 3.36466018e-01
- 3.51539812e-01 3.66763093e-01 3.82120532e-01 3.97596666e-01
- 4.13175911e-01 4.28842581e-01 4.44580900e-01 4.60375022e-01
- 4.76209042e-01 4.92067018e-01 5.07932982e-01 5.23790958e-01
- 5.39624978e-01 5.55419100e-01 5.71157419e-01 5.86824089e-01
- 6.02403334e-01 6.17879468e-01 6.33236907e-01 6.48460188e-01
- 6.63533982e-01 6.78443111e-01 6.93172563e-01 7.07707507e-01
- 7.22033306e-01 7.36135537e-01 7.50000000e-01 7.63612734e-01
- 7.76960032e-01 7.90028455e-01 8.02804844e-01 8.15276334e-01
- 8.27430367e-01 8.39254706e-01 8.50737444e-01 8.61867019e-01
- 8.72632225e-01 8.83022222e-01 8.93026547e-01 9.02635129e-01
- 9.11838291e-01 9.20626766e-01 9.28991707e-01 9.36924689e-01
- 9.44417724e-01 9.51463269e-01 9.58054229e-01 9.64183967e-01
- 9.69846310e-01 9.75035559e-01 9.79746487e-01 9.83974351e-01
- 9.87714893e-01 9.90964349e-01 9.93719444e-01 9.95977406e-01
- 9.97735961e-01 9.98993338e-01 9.99748271e-01 1.00000000e+00
- 9.99748271e-01 9.98993338e-01 9.97735961e-01 9.95977406e-01
- 9.93719444e-01 9.90964349e-01 9.87714893e-01 9.83974351e-01
- 9.79746487e-01 9.75035559e-01 9.69846310e-01 9.64183967e-01
- 9.58054229e-01 9.51463269e-01 9.44417724e-01 9.36924689e-01
- 9.28991707e-01 9.20626766e-01 9.11838291e-01 9.02635129e-01
- 8.93026547e-01 8.83022222e-01 8.72632225e-01 8.61867019e-01
- 8.50737444e-01 8.39254706e-01 8.27430367e-01 8.15276334e-01
- 8.02804844e-01 7.90028455e-01 7.76960032e-01 7.63612734e-01
- 7.50000000e-01 7.36135537e-01 7.22033306e-01 7.07707507e-01
- 6.93172563e-01 6.78443111e-01 6.63533982e-01 6.48460188e-01
- 6.33236907e-01 6.17879468e-01 6.02403334e-01 5.86824089e-01
- 5.71157419e-01 5.55419100e-01 5.39624978e-01 5.23790958e-01
- 5.07932982e-01 4.92067018e-01 4.76209042e-01 4.60375022e-01
- 4.44580900e-01 4.28842581e-01 4.13175911e-01 3.97596666e-01
- 3.82120532e-01 3.66763093e-01 3.51539812e-01 3.36466018e-01
- 3.21556889e-01 3.06827437e-01 2.92292493e-01 2.77966694e-01
- 2.63864463e-01 2.50000000e-01 2.36387266e-01 2.23039968e-01
- 2.09971545e-01 1.97195156e-01 1.84723666e-01 1.72569633e-01
- 1.60745294e-01 1.49262556e-01 1.38132981e-01 1.27367775e-01
- 1.16977778e-01 1.06973453e-01 9.73648712e-02 8.81617093e-02
- 7.93732336e-02 7.10082934e-02 6.30753115e-02 5.55822757e-02
- 4.85367309e-02 4.19457713e-02 3.58160335e-02 3.01536896e-02
- 2.49644411e-02 2.02535132e-02 1.60256493e-02 1.22851066e-02
- 9.03565137e-03 6.28055566e-03 4.02259358e-03 2.26403871e-03
- 1.00666176e-03 2.51728808e-04]  [ 0.          0.00961446  0.02009799  0.03068111  0.04113909  0.05172203
-  0.06196484  0.07232509  0.08237074  0.09212461  0.1014782   0.11108724
-  0.12119772  0.13128298  0.14115052  0.1508826   0.16039743  0.16965594
-  0.17873706  0.1875441   0.19605555  0.20411389  0.21173694  0.21902442
-  0.22562364  0.23190658  0.23784273  0.24344245  0.24870201  0.25359838
-  0.25816122  0.26235769  0.2662825   0.26991069  0.27325294  0.27629426
-  0.27905322  0.28152161  0.28372148  0.28562123  0.28720139  0.28847286
-  0.28943224  0.29009618  0.29048917  0.290573    0.29038695  0.28993786
-  0.2891893   0.28818705  0.28692373  0.28539624  0.28358942  0.28159096
-  0.27939157  0.27699898  0.27448937  0.27186734  0.26912464  0.26629024
-  0.26338009  0.26040408  0.25738024  0.25432356  0.25123857  0.24814729
-  0.24506615  0.24200867  0.23898794  0.23600195  0.23303817  0.23014754
-  0.22732341  0.22458865  0.22196988  0.21942333  0.21697534  0.21463751
-  0.21240975  0.210283    0.2082865   0.20640118  0.20464098  0.20296971
-  0.20138853  0.19990288  0.19851648  0.1972271   0.19604019  0.1949542
-  0.19396831  0.19301665  0.19210009  0.19130387  0.19062878  0.19007551
-  0.1896446   0.18688237  0.18369077  0.18262654  0.18258324  0.1824534
-  0.18223713  0.18193466  0.18154629  0.18107242  0.18051352  0.17986604
-  0.17902221  0.17808197  0.17704625  0.17564539  0.17400006  0.17217466
-  0.16978469  0.16724293  0.1640559   0.16064317  0.15662296  0.15222742
-  0.14730191  0.14169521  0.13554458  0.12878762  0.12117948  0.11287882
-  0.10383363  0.0940176   0.08345467  0.07217951  0.06022695  0.0476921
-  0.03465795  0.02119749  0.00738003 -0.00663088 -0.02076815 -0.03492629
- -0.0489827  -0.06294912 -0.07670871 -0.08989053 -0.10267547 -0.11501216
- -0.12683036 -0.13818345 -0.14900076 -0.15920936 -0.16886297 -0.17804687
- -0.18678976 -0.19505431 -0.20284597 -0.21008893 -0.21677911 -0.22298947
- -0.228554   -0.23347062 -0.23769467 -0.24108136 -0.24370266 -0.24552406
- -0.24656618 -0.24680488 -0.24634337 -0.24523001 -0.24353847 -0.24130027
- -0.23842982 -0.23507222 -0.23124411 -0.22672126 -0.22168884 -0.21615481
- -0.2100415  -0.20355375 -0.19678786 -0.18965264 -0.18230108 -0.1746862
- -0.16688459 -0.15888574 -0.15083294 -0.14313377 -0.13509555 -0.12656456
- -0.11777199 -0.10875138 -0.09957619 -0.09016215 -0.08034882 -0.07037605
- -0.06066991 -0.05105958 -0.04045763 -0.02812167 -0.01791582 -0.00815034] [3440000000.0, 41800000000.0, 27700000000.0, 13600000000.0, 200000000000.0, 205000000000.0, 118000000000.0, 98460000.0, 114500000000.0, 1000000.0, 4560000.0] [3440000000.0, 14000000000.0, 13650000000.0, 13300000000.0, 200000000000.0, 205000000000.0, 118000000000.0, 98460000.0, 8390000000.0, 1000000.0, 4560000.0] [1323000000.0, 2630000000.0, 7200000000.0, 11800000000.0, 79300000000.0, 80000000000.0, 47600000000.0, 98460000.0, 5990000000.0, 384615.3846153846, 1520000.0] [0.3, 0.28, 0.39, 0.49, 0.3, 0.3, 0.3, 0.3, 0.27, 0.3, 0.49] [1235.0, 1920.0, 1850.0, 1780.0, 7800.0, 7850.0, 7200.0, 200.0, 1220.0, 1150.0, 1100.0] [0.         0.45054387 0.59760341 0.90060069 1.        ] [5. 5. 5. 5.] [1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1.] [5.00000000e-05 2.82000000e-03 1.07726981e-03 2.00000000e-02
- 2.82000000e-03 5.00000000e-05 2.82000000e-03 1.07726981e-03
- 4.23000000e-02 2.82000000e-03 5.00000000e-05 2.82000000e-03
- 1.07726981e-03 9.00000000e-02 2.82000000e-03 5.00000000e-05
- 2.82000000e-03 1.07726981e-03 7.05000000e-03 2.82000000e-03] [0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0.] [1 3 3 8 3 1 3 3 9 3 1 3 3 8 3 1 3 3 2 3] [0.         0.22176114 0.39356337 0.89603809 1.        ] [5. 5. 5. 5.] [1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1.] [5.00000000e-05 2.82000000e-03 1.07726981e-03 2.00000000e-02
- 2.82000000e-03 5.00000000e-05 2.82000000e-03 1.07726981e-03
- 4.23000000e-02 2.82000000e-03 5.00000000e-05 2.82000000e-03
- 1.07726981e-03 9.00000000e-02 2.82000000e-03 5.00000000e-05
- 2.82000000e-03 1.07726981e-03 7.05000000e-03 2.82000000e-03] [0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0.] [1 3 3 8 3 1 3 3 9 3 1 3 3 8 3 1 3 3 2 3] 2 [0.4660247  0.58448794] [3. 3.] [1. 1. 1. 1. 1. 1.] [0.002 0.04  0.002 0.002 0.04  0.002] [0. 0. 0. 0. 0. 0.] [4 8 4 4 8 4]
-
-
-5 4.5705983562554815 0.0 0.0 0.4 [0.00000000e+00 2.51728808e-04 1.00666176e-03 2.26403871e-03
- 4.02259358e-03 6.28055566e-03 9.03565137e-03 1.22851066e-02
- 1.60256493e-02 2.02535132e-02 2.49644411e-02 3.01536896e-02
- 3.58160335e-02 4.19457713e-02 4.85367309e-02 5.55822757e-02
- 6.30753115e-02 7.10082934e-02 7.93732336e-02 8.81617093e-02
- 9.73648712e-02 1.06973453e-01 1.16977778e-01 1.27367775e-01
- 1.38132981e-01 1.49262556e-01 1.60745294e-01 1.72569633e-01
- 1.84723666e-01 1.97195156e-01 2.09971545e-01 2.23039968e-01
- 2.36387266e-01 2.50000000e-01 2.63864463e-01 2.77966694e-01
- 2.92292493e-01 3.06827437e-01 3.21556889e-01 3.36466018e-01
- 3.51539812e-01 3.66763093e-01 3.82120532e-01 3.97596666e-01
- 4.13175911e-01 4.28842581e-01 4.44580900e-01 4.60375022e-01
- 4.76209042e-01 4.92067018e-01 5.07932982e-01 5.23790958e-01
- 5.39624978e-01 5.55419100e-01 5.71157419e-01 5.86824089e-01
- 6.02403334e-01 6.17879468e-01 6.33236907e-01 6.48460188e-01
- 6.63533982e-01 6.78443111e-01 6.93172563e-01 7.07707507e-01
- 7.22033306e-01 7.36135537e-01 7.50000000e-01 7.63612734e-01
- 7.76960032e-01 7.90028455e-01 8.02804844e-01 8.15276334e-01
- 8.27430367e-01 8.39254706e-01 8.50737444e-01 8.61867019e-01
- 8.72632225e-01 8.83022222e-01 8.93026547e-01 9.02635129e-01
- 9.11838291e-01 9.20626766e-01 9.28991707e-01 9.36924689e-01
- 9.44417724e-01 9.51463269e-01 9.58054229e-01 9.64183967e-01
- 9.69846310e-01 9.75035559e-01 9.79746487e-01 9.83974351e-01
- 9.87714893e-01 9.90964349e-01 9.93719444e-01 9.95977406e-01
- 9.97735961e-01 9.98993338e-01 9.99748271e-01 1.00000000e+00
- 9.99748271e-01 9.98993338e-01 9.97735961e-01 9.95977406e-01
- 9.93719444e-01 9.90964349e-01 9.87714893e-01 9.83974351e-01
- 9.79746487e-01 9.75035559e-01 9.69846310e-01 9.64183967e-01
- 9.58054229e-01 9.51463269e-01 9.44417724e-01 9.36924689e-01
- 9.28991707e-01 9.20626766e-01 9.11838291e-01 9.02635129e-01
- 8.93026547e-01 8.83022222e-01 8.72632225e-01 8.61867019e-01
- 8.50737444e-01 8.39254706e-01 8.27430367e-01 8.15276334e-01
- 8.02804844e-01 7.90028455e-01 7.76960032e-01 7.63612734e-01
- 7.50000000e-01 7.36135537e-01 7.22033306e-01 7.07707507e-01
- 6.93172563e-01 6.78443111e-01 6.63533982e-01 6.48460188e-01
- 6.33236907e-01 6.17879468e-01 6.02403334e-01 5.86824089e-01
- 5.71157419e-01 5.55419100e-01 5.39624978e-01 5.23790958e-01
- 5.07932982e-01 4.92067018e-01 4.76209042e-01 4.60375022e-01
- 4.44580900e-01 4.28842581e-01 4.13175911e-01 3.97596666e-01
- 3.82120532e-01 3.66763093e-01 3.51539812e-01 3.36466018e-01
- 3.21556889e-01 3.06827437e-01 2.92292493e-01 2.77966694e-01
- 2.63864463e-01 2.50000000e-01 2.36387266e-01 2.23039968e-01
- 2.09971545e-01 1.97195156e-01 1.84723666e-01 1.72569633e-01
- 1.60745294e-01 1.49262556e-01 1.38132981e-01 1.27367775e-01
- 1.16977778e-01 1.06973453e-01 9.73648712e-02 8.81617093e-02
- 7.93732336e-02 7.10082934e-02 6.30753115e-02 5.55822757e-02
- 4.85367309e-02 4.19457713e-02 3.58160335e-02 3.01536896e-02
- 2.49644411e-02 2.02535132e-02 1.60256493e-02 1.22851066e-02
- 9.03565137e-03 6.28055566e-03 4.02259358e-03 2.26403871e-03
- 1.00666176e-03 2.51728808e-04]  [ 0.          0.00564708  0.01287509  0.02079393  0.02893067  0.03732045
-  0.04570963  0.05411059  0.06249739  0.07090825  0.07924073  0.08756286
-  0.09570151  0.10377924  0.11170708  0.11949121  0.12716077  0.13462516
-  0.14190438  0.14904058  0.15595542  0.1626544   0.16914081  0.17539936
-  0.1813675   0.18708012  0.19252037  0.19767446  0.2024734   0.20697389
-  0.21115827  0.21502532  0.21859097  0.22185037  0.22484839  0.22758883
-  0.23008885  0.23235188  0.23438898  0.23618514  0.23775212  0.23909332
-  0.2402132   0.24112099  0.24181519  0.24231092  0.24261725  0.2427208
-  0.24264181  0.2423971   0.24198707  0.2414381   0.24077192  0.23998709
-  0.23909872  0.2381382   0.23710398  0.23600273  0.23484606  0.23364503
-  0.23240844  0.23114596  0.22986528  0.22857565  0.22728388  0.22599571
-  0.22471761  0.22345655  0.22221294  0.22099158  0.21979666  0.21863317
-  0.21751645  0.21644351  0.21541426  0.21443221  0.21349386  0.21261421
-  0.21177434  0.21098907  0.21024839  0.20956419  0.20891379  0.2082991
-  0.20771851  0.2071665   0.20664995  0.20616853  0.20572134  0.2053115
-  0.20493945  0.2046028   0.20430448  0.20404533  0.2038256   0.20364552
-  0.20350527  0.20081148  0.1976129   0.19654635  0.19653828  0.19651409
-  0.19647379  0.19641744  0.19634508  0.19625678  0.19615265  0.19603278
-  0.19577484  0.19524194  0.19465492  0.19401439  0.1929522   0.19166725
-  0.19029368  0.18818057  0.18591336  0.18319719  0.18015854  0.17667821
-  0.17285382  0.16847903  0.16376383  0.15840227  0.15256875  0.14626776
-  0.13935839  0.13188861  0.12390827  0.11540727  0.10639007  0.09688622
-  0.08690209  0.07648242  0.06567519  0.05450427  0.0430056   0.03122944
-  0.01921805  0.00701482 -0.00532589 -0.0177448  -0.03019424 -0.04258222
- -0.05482268 -0.06686771 -0.07863566 -0.09001883 -0.10097862 -0.11146909
- -0.12146866 -0.13087109 -0.13968572 -0.1478927  -0.15546102 -0.1622476
- -0.16831579 -0.17367112 -0.17825612 -0.18197933 -0.18497124 -0.18720625
- -0.18873185 -0.18948047 -0.18961367 -0.18918944 -0.18822661 -0.18672586
- -0.18480846 -0.18247101 -0.1796697  -0.17650088 -0.17298931 -0.16908484
- -0.16489104 -0.16036799 -0.1555536  -0.15050474 -0.14513386 -0.13958088
- -0.1337919  -0.12784274 -0.12170732 -0.11546424 -0.10906704 -0.10255531
- -0.09582731 -0.08899525 -0.0819927  -0.0748683  -0.06755623 -0.05981473
- -0.05201741 -0.04456567 -0.03708454 -0.02879087 -0.01857376 -0.00887937] [3440000000.0, 41800000000.0, 27700000000.0, 13600000000.0, 200000000000.0, 205000000000.0, 118000000000.0, 98460000.0, 114500000000.0, 1000000.0, 4560000.0] [3440000000.0, 14000000000.0, 13650000000.0, 13300000000.0, 200000000000.0, 205000000000.0, 118000000000.0, 98460000.0, 8390000000.0, 1000000.0, 4560000.0] [1323000000.0, 2630000000.0, 7200000000.0, 11800000000.0, 79300000000.0, 80000000000.0, 47600000000.0, 98460000.0, 5990000000.0, 384615.3846153846, 1520000.0] [0.3, 0.28, 0.39, 0.49, 0.3, 0.3, 0.3, 0.3, 0.27, 0.3, 0.49] [1235.0, 1920.0, 1850.0, 1780.0, 7800.0, 7850.0, 7200.0, 200.0, 1220.0, 1150.0, 1100.0] [0.         0.41489445 0.55692304 0.90195151 1.        ] [4. 4. 4. 4.] [1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1.] [5.00e-05 2.82e-03 2.00e-02 2.82e-03 5.00e-05 2.82e-03 4.23e-02 2.82e-03
- 5.00e-05 2.82e-03 9.00e-02 2.82e-03 5.00e-05 2.82e-03 7.05e-03 2.82e-03] [0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0.] [1 3 8 3 1 3 9 3 1 3 8 3 1 3 2 3] [0.         0.23220727 0.39812822 0.89795416 1.        ] [4. 4. 4. 4.] [1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1.] [5.00e-05 2.82e-03 2.00e-02 2.82e-03 5.00e-05 2.82e-03 4.23e-02 2.82e-03
- 5.00e-05 2.82e-03 9.00e-02 2.82e-03 5.00e-05 2.82e-03 7.05e-03 2.82e-03] [0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0.] [1 3 8 3 1 3 9 3 1 3 8 3 1 3 2 3] 2 [0.42988073 0.54403751] [3. 3.] [1. 1. 1. 1. 1. 1.] [0.002 0.04  0.002 0.002 0.04  0.002] [0. 0. 0. 0. 0. 0.] [4 8 4 4 8 4]
-
-
-6 4.6343256119198495 0.0 0.0 0.4 [0.00000000e+00 2.51728808e-04 1.00666176e-03 2.26403871e-03
- 4.02259358e-03 6.28055566e-03 9.03565137e-03 1.22851066e-02
- 1.60256493e-02 2.02535132e-02 2.49644411e-02 3.01536896e-02
- 3.58160335e-02 4.19457713e-02 4.85367309e-02 5.55822757e-02
- 6.30753115e-02 7.10082934e-02 7.93732336e-02 8.81617093e-02
- 9.73648712e-02 1.06973453e-01 1.16977778e-01 1.27367775e-01
- 1.38132981e-01 1.49262556e-01 1.60745294e-01 1.72569633e-01
- 1.84723666e-01 1.97195156e-01 2.09971545e-01 2.23039968e-01
- 2.36387266e-01 2.50000000e-01 2.63864463e-01 2.77966694e-01
- 2.92292493e-01 3.06827437e-01 3.21556889e-01 3.36466018e-01
- 3.51539812e-01 3.66763093e-01 3.82120532e-01 3.97596666e-01
- 4.13175911e-01 4.28842581e-01 4.44580900e-01 4.60375022e-01
- 4.76209042e-01 4.92067018e-01 5.07932982e-01 5.23790958e-01
- 5.39624978e-01 5.55419100e-01 5.71157419e-01 5.86824089e-01
- 6.02403334e-01 6.17879468e-01 6.33236907e-01 6.48460188e-01
- 6.63533982e-01 6.78443111e-01 6.93172563e-01 7.07707507e-01
- 7.22033306e-01 7.36135537e-01 7.50000000e-01 7.63612734e-01
- 7.76960032e-01 7.90028455e-01 8.02804844e-01 8.15276334e-01
- 8.27430367e-01 8.39254706e-01 8.50737444e-01 8.61867019e-01
- 8.72632225e-01 8.83022222e-01 8.93026547e-01 9.02635129e-01
- 9.11838291e-01 9.20626766e-01 9.28991707e-01 9.36924689e-01
- 9.44417724e-01 9.51463269e-01 9.58054229e-01 9.64183967e-01
- 9.69846310e-01 9.75035559e-01 9.79746487e-01 9.83974351e-01
- 9.87714893e-01 9.90964349e-01 9.93719444e-01 9.95977406e-01
- 9.97735961e-01 9.98993338e-01 9.99748271e-01 1.00000000e+00
- 9.99748271e-01 9.98993338e-01 9.97735961e-01 9.95977406e-01
- 9.93719444e-01 9.90964349e-01 9.87714893e-01 9.83974351e-01
- 9.79746487e-01 9.75035559e-01 9.69846310e-01 9.64183967e-01
- 9.58054229e-01 9.51463269e-01 9.44417724e-01 9.36924689e-01
- 9.28991707e-01 9.20626766e-01 9.11838291e-01 9.02635129e-01
- 8.93026547e-01 8.83022222e-01 8.72632225e-01 8.61867019e-01
- 8.50737444e-01 8.39254706e-01 8.27430367e-01 8.15276334e-01
- 8.02804844e-01 7.90028455e-01 7.76960032e-01 7.63612734e-01
- 7.50000000e-01 7.36135537e-01 7.22033306e-01 7.07707507e-01
- 6.93172563e-01 6.78443111e-01 6.63533982e-01 6.48460188e-01
- 6.33236907e-01 6.17879468e-01 6.02403334e-01 5.86824089e-01
- 5.71157419e-01 5.55419100e-01 5.39624978e-01 5.23790958e-01
- 5.07932982e-01 4.92067018e-01 4.76209042e-01 4.60375022e-01
- 4.44580900e-01 4.28842581e-01 4.13175911e-01 3.97596666e-01
- 3.82120532e-01 3.66763093e-01 3.51539812e-01 3.36466018e-01
- 3.21556889e-01 3.06827437e-01 2.92292493e-01 2.77966694e-01
- 2.63864463e-01 2.50000000e-01 2.36387266e-01 2.23039968e-01
- 2.09971545e-01 1.97195156e-01 1.84723666e-01 1.72569633e-01
- 1.60745294e-01 1.49262556e-01 1.38132981e-01 1.27367775e-01
- 1.16977778e-01 1.06973453e-01 9.73648712e-02 8.81617093e-02
- 7.93732336e-02 7.10082934e-02 6.30753115e-02 5.55822757e-02
- 4.85367309e-02 4.19457713e-02 3.58160335e-02 3.01536896e-02
- 2.49644411e-02 2.02535132e-02 1.60256493e-02 1.22851066e-02
- 9.03565137e-03 6.28055566e-03 4.02259358e-03 2.26403871e-03
- 1.00666176e-03 2.51728808e-04] [ 0.          0.00589183  0.01268659  0.01923175  0.0262915   0.03350608
-  0.0409496   0.04847493  0.05602286  0.0636605   0.07126877  0.07893115
-  0.0864394   0.09383399  0.10106507  0.10825266  0.11532127  0.12227493
-  0.12911074  0.13575336  0.14227825  0.14863212  0.15477157  0.16074067
-  0.16647248  0.17195683  0.17721017  0.18218915  0.18684185  0.19119905
-  0.19524096  0.1989241   0.20226083  0.20531732  0.20811456  0.21067168
-  0.21298953  0.21508693  0.21700456  0.21875132  0.22031992  0.22170298
-  0.22290679  0.22389781  0.22472996  0.22539791  0.22591038  0.22626962
-  0.22648554  0.22655932  0.22649025  0.22629969  0.22600389  0.22560999
-  0.22512443  0.224555    0.22391472  0.22320617  0.22243842  0.22162778
-  0.22078047  0.21989899  0.21899365  0.21806782  0.21712602  0.21620154
-  0.21530151  0.21441533  0.21353795  0.21266257  0.2117959   0.21094306
-  0.21011226  0.20930762  0.2085321   0.20778659  0.20706934  0.20638976
-  0.20573836  0.20511708  0.20453054  0.20397868  0.20345303  0.202954
-  0.2024807   0.20201636  0.20158198  0.20117363  0.20078882  0.20043616
-  0.20011601  0.1998192   0.1995552   0.19932587  0.19913142  0.19897206
-  0.19884795  0.19728519  0.19383306  0.19268196  0.19268072  0.19267701
-  0.19267084  0.19266221  0.19265112  0.19263759  0.19262163  0.19260327
-  0.19250229  0.19208073  0.19161637  0.19110968  0.19027438  0.18911833
-  0.18788256  0.18605921  0.18391921  0.18149157  0.17862491  0.17548766
-  0.17186917  0.16790253  0.16350739  0.15858977  0.15329643  0.14743703
-  0.14111439  0.13435723  0.12713139  0.11939257  0.11124025  0.10269077
-  0.09376402  0.08448138  0.07487314  0.06496286  0.05477383  0.0443371
-  0.0336955   0.02283996  0.01181626  0.00064508 -0.01062984 -0.02195355
- -0.03326094 -0.04451225 -0.05564487 -0.06660097 -0.07732204 -0.08774352
- -0.09775029 -0.10732904 -0.1164318  -0.12498126 -0.13282247 -0.13997485
- -0.14640689 -0.15208504 -0.15684301 -0.16078489 -0.16390917 -0.16623704
- -0.1676836  -0.16838012 -0.16842575 -0.16786868 -0.166704   -0.16507411
- -0.16302133 -0.16052623 -0.15766525 -0.15447663 -0.15089515 -0.14704121
- -0.1429074  -0.1385078  -0.13389434 -0.12904646 -0.12402824 -0.11880174
- -0.11345749 -0.10796128 -0.10239796 -0.09673686 -0.09103304 -0.08518077
- -0.07924933 -0.07329566 -0.06723758 -0.06117004 -0.05505616 -0.04880743
- -0.04244291 -0.03614832 -0.03001623 -0.02383718 -0.01727108 -0.00958906] [3440000000.0, 41800000000.0, 27700000000.0, 13600000000.0, 200000000000.0, 205000000000.0, 118000000000.0, 98460000.0, 114500000000.0, 1000000.0, 4560000.0] [3440000000.0, 14000000000.0, 13650000000.0, 13300000000.0, 200000000000.0, 205000000000.0, 118000000000.0, 98460000.0, 8390000000.0, 1000000.0, 4560000.0] [1323000000.0, 2630000000.0, 7200000000.0, 11800000000.0, 79300000000.0, 80000000000.0, 47600000000.0, 98460000.0, 5990000000.0, 384615.3846153846, 1520000.0] [0.3, 0.28, 0.39, 0.49, 0.3, 0.3, 0.3, 0.3, 0.27, 0.3, 0.49] [1235.0, 1920.0, 1850.0, 1780.0, 7800.0, 7850.0, 7200.0, 200.0, 1220.0, 1150.0, 1100.0] [0.         0.40319807 0.54446337 0.90298863 1.        ] [4. 4. 4. 4.] [1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1.] [5.00e-05 2.82e-03 2.00e-02 2.82e-03 5.00e-05 2.82e-03 4.23e-02 2.82e-03
- 5.00e-05 2.82e-03 9.00e-02 2.82e-03 5.00e-05 2.82e-03 7.05e-03 2.82e-03] [0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0.] [1 3 8 3 1 3 9 3 1 3 8 3 1 3 2 3] [0.         0.24893251 0.41282855 0.8997402  1.        ] [4. 4. 4. 4.] [1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1.] [5.00e-05 2.82e-03 2.00e-02 2.82e-03 5.00e-05 2.82e-03 4.23e-02 2.82e-03
- 5.00e-05 2.82e-03 9.00e-02 2.82e-03 5.00e-05 2.82e-03 7.05e-03 2.82e-03] [0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0.] [1 3 8 3 1 3 9 3 1 3 8 3 1 3 2 3] 2 [0.41790419 0.53154924] [3. 3.] [1. 1. 1. 1. 1. 1.] [0.002 0.04  0.002 0.002 0.04  0.002] [0. 0. 0. 0. 0. 0.] [4 8 4 4 8 4]
-
-
-7 4.646781924312676 0.0 0.0 0.4 [0.00000000e+00 2.51728808e-04 1.00666176e-03 2.26403871e-03
- 4.02259358e-03 6.28055566e-03 9.03565137e-03 1.22851066e-02
- 1.60256493e-02 2.02535132e-02 2.49644411e-02 3.01536896e-02
- 3.58160335e-02 4.19457713e-02 4.85367309e-02 5.55822757e-02
- 6.30753115e-02 7.10082934e-02 7.93732336e-02 8.81617093e-02
- 9.73648712e-02 1.06973453e-01 1.16977778e-01 1.27367775e-01
- 1.38132981e-01 1.49262556e-01 1.60745294e-01 1.72569633e-01
- 1.84723666e-01 1.97195156e-01 2.09971545e-01 2.23039968e-01
- 2.36387266e-01 2.50000000e-01 2.63864463e-01 2.77966694e-01
- 2.92292493e-01 3.06827437e-01 3.21556889e-01 3.36466018e-01
- 3.51539812e-01 3.66763093e-01 3.82120532e-01 3.97596666e-01
- 4.13175911e-01 4.28842581e-01 4.44580900e-01 4.60375022e-01
- 4.76209042e-01 4.92067018e-01 5.07932982e-01 5.23790958e-01
- 5.39624978e-01 5.55419100e-01 5.71157419e-01 5.86824089e-01
- 6.02403334e-01 6.17879468e-01 6.33236907e-01 6.48460188e-01
- 6.63533982e-01 6.78443111e-01 6.93172563e-01 7.07707507e-01
- 7.22033306e-01 7.36135537e-01 7.50000000e-01 7.63612734e-01
- 7.76960032e-01 7.90028455e-01 8.02804844e-01 8.15276334e-01
- 8.27430367e-01 8.39254706e-01 8.50737444e-01 8.61867019e-01
- 8.72632225e-01 8.83022222e-01 8.93026547e-01 9.02635129e-01
- 9.11838291e-01 9.20626766e-01 9.28991707e-01 9.36924689e-01
- 9.44417724e-01 9.51463269e-01 9.58054229e-01 9.64183967e-01
- 9.69846310e-01 9.75035559e-01 9.79746487e-01 9.83974351e-01
- 9.87714893e-01 9.90964349e-01 9.93719444e-01 9.95977406e-01
- 9.97735961e-01 9.98993338e-01 9.99748271e-01 1.00000000e+00
- 9.99748271e-01 9.98993338e-01 9.97735961e-01 9.95977406e-01
- 9.93719444e-01 9.90964349e-01 9.87714893e-01 9.83974351e-01
- 9.79746487e-01 9.75035559e-01 9.69846310e-01 9.64183967e-01
- 9.58054229e-01 9.51463269e-01 9.44417724e-01 9.36924689e-01
- 9.28991707e-01 9.20626766e-01 9.11838291e-01 9.02635129e-01
- 8.93026547e-01 8.83022222e-01 8.72632225e-01 8.61867019e-01
- 8.50737444e-01 8.39254706e-01 8.27430367e-01 8.15276334e-01
- 8.02804844e-01 7.90028455e-01 7.76960032e-01 7.63612734e-01
- 7.50000000e-01 7.36135537e-01 7.22033306e-01 7.07707507e-01
- 6.93172563e-01 6.78443111e-01 6.63533982e-01 6.48460188e-01
- 6.33236907e-01 6.17879468e-01 6.02403334e-01 5.86824089e-01
- 5.71157419e-01 5.55419100e-01 5.39624978e-01 5.23790958e-01
- 5.07932982e-01 4.92067018e-01 4.76209042e-01 4.60375022e-01
- 4.44580900e-01 4.28842581e-01 4.13175911e-01 3.97596666e-01
- 3.82120532e-01 3.66763093e-01 3.51539812e-01 3.36466018e-01
- 3.21556889e-01 3.06827437e-01 2.92292493e-01 2.77966694e-01
- 2.63864463e-01 2.50000000e-01 2.36387266e-01 2.23039968e-01
- 2.09971545e-01 1.97195156e-01 1.84723666e-01 1.72569633e-01
- 1.60745294e-01 1.49262556e-01 1.38132981e-01 1.27367775e-01
- 1.16977778e-01 1.06973453e-01 9.73648712e-02 8.81617093e-02
- 7.93732336e-02 7.10082934e-02 6.30753115e-02 5.55822757e-02
- 4.85367309e-02 4.19457713e-02 3.58160335e-02 3.01536896e-02
- 2.49644411e-02 2.02535132e-02 1.60256493e-02 1.22851066e-02
- 9.03565137e-03 6.28055566e-03 4.02259358e-03 2.26403871e-03
- 1.00666176e-03 2.51728808e-04]  [ 0.          0.00746219  0.01396329  0.01992013  0.02615354  0.03247978
-  0.03906494  0.04576265  0.05251823  0.05938311  0.06624734  0.07319394
-  0.08009732  0.08704283  0.09396398  0.10081345  0.10759241  0.11427433
-  0.12084545  0.12727866  0.13358273  0.13975495  0.14572913  0.15155195
-  0.15713582  0.16252254  0.16770502  0.17257301  0.17719741  0.18152953
-  0.1855071   0.18914606  0.19248981  0.19555658  0.19833454  0.20087534
-  0.20320523  0.20533461  0.20729911  0.20907159  0.21066781  0.21208656
-  0.21332295  0.21438607  0.21528865  0.21603082  0.21663329  0.21708463
-  0.21740068  0.21756973  0.2176076   0.21752938  0.21734889  0.21707006
-  0.21669425  0.21623153  0.21569472  0.21507661  0.21440021  0.21367471
-  0.21290241  0.21209209  0.21125202  0.21038654  0.20949323  0.20856122
-  0.20761637  0.20667293  0.20574248  0.20480283  0.2038761   0.20296535
-  0.20207202  0.20120792  0.20037178  0.19956387  0.19878681  0.19804404
-  0.19733029  0.19664573  0.19599818  0.19538367  0.19479756  0.19424083
-  0.19370712  0.19317597  0.1926791   0.19220756  0.19176033  0.19135046
-  0.19097838  0.19062733  0.19031487  0.19004343  0.18981329  0.18962468
-  0.18947778  0.18884467  0.18506655  0.18380675  0.18381073  0.18382265
-  0.1838425   0.18387027  0.18390592  0.18394942  0.18400073  0.18405979
-  0.18408129  0.18376151  0.18340927  0.18302491  0.18239328  0.18137111
-  0.18027845  0.17874732  0.17679324  0.17466794  0.17201479  0.16922209
-  0.16584339  0.16226471  0.15814448  0.15368508  0.14877755  0.14337212
-  0.13760066  0.13137875  0.12467674  0.11760117  0.11015005  0.10232194
-  0.09411156  0.08558743  0.07676952  0.06767606  0.05832659  0.04874387
-  0.0389619   0.02900202  0.01887072  0.00860606 -0.00176943 -0.01221809
- -0.02268658 -0.03312272 -0.04350139 -0.05376423 -0.06386155 -0.07375114
- -0.08336156 -0.09263761 -0.10147472 -0.10987274 -0.11773152 -0.12498453
- -0.1315009  -0.13726294 -0.14227136 -0.14646563 -0.14976136 -0.15218646
- -0.15377915 -0.1546084  -0.15463941 -0.15397319 -0.15273995 -0.15102797
- -0.14878372 -0.14616632 -0.14320323 -0.13985485 -0.13622364 -0.13228306
- -0.12807726 -0.12365733 -0.11900427 -0.11418061 -0.10918609 -0.1040424
- -0.09879823 -0.09343606 -0.08803322 -0.082571   -0.0771146  -0.07170525
- -0.06631973 -0.06094434 -0.05553664 -0.05011262 -0.04469635 -0.03922073
- -0.03367385 -0.02810433 -0.02264161 -0.01724888 -0.01182505 -0.00631792] [3440000000.0, 41800000000.0, 27700000000.0, 13600000000.0, 200000000000.0, 205000000000.0, 118000000000.0, 98460000.0, 114500000000.0, 1000000.0, 4560000.0] [3440000000.0, 14000000000.0, 13650000000.0, 13300000000.0, 200000000000.0, 205000000000.0, 118000000000.0, 98460000.0, 8390000000.0, 1000000.0, 4560000.0] [1323000000.0, 2630000000.0, 7200000000.0, 11800000000.0, 79300000000.0, 80000000000.0, 47600000000.0, 98460000.0, 5990000000.0, 384615.3846153846, 1520000.0] [0.3, 0.28, 0.39, 0.49, 0.3, 0.3, 0.3, 0.3, 0.27, 0.3, 0.49] [1235.0, 1920.0, 1850.0, 1780.0, 7800.0, 7850.0, 7200.0, 200.0, 1220.0, 1150.0, 1100.0] [0.         0.39263083 0.53541334 0.90310294 1.        ] [4. 4. 4. 4.] [1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1.] [5.00e-05 2.82e-03 2.00e-02 2.82e-03 5.00e-05 2.82e-03 4.23e-02 2.82e-03
- 5.00e-05 2.82e-03 9.00e-02 2.82e-03 5.00e-05 2.82e-03 7.05e-03 2.82e-03] [0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0.] [1 3 8 3 1 3 9 3 1 3 8 3 1 3 2 3] [0.         0.25985429 0.42300587 0.90058236 1.        ] [4. 4. 4. 4.] [1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1.] [5.00e-05 2.82e-03 2.00e-02 2.82e-03 5.00e-05 2.82e-03 4.23e-02 2.82e-03
- 5.00e-05 2.82e-03 9.00e-02 2.82e-03 5.00e-05 2.82e-03 7.05e-03 2.82e-03] [0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0.] [1 3 8 3 1 3 9 3 1 3 8 3 1 3 2 3] 2 [0.40733145 0.52230947] [3. 3.] [1. 1. 1. 1. 1. 1.] [0.002 0.04  0.002 0.002 0.04  0.002] [0. 0. 0. 0. 0. 0.] [4 8 4 4 8 4]
-
-
-8 4.545529346958071 0.0 0.0 0.4 [0.00000000e+00 2.51728808e-04 1.00666176e-03 2.26403871e-03
- 4.02259358e-03 6.28055566e-03 9.03565137e-03 1.22851066e-02
- 1.60256493e-02 2.02535132e-02 2.49644411e-02 3.01536896e-02
- 3.58160335e-02 4.19457713e-02 4.85367309e-02 5.55822757e-02
- 6.30753115e-02 7.10082934e-02 7.93732336e-02 8.81617093e-02
- 9.73648712e-02 1.06973453e-01 1.16977778e-01 1.27367775e-01
- 1.38132981e-01 1.49262556e-01 1.60745294e-01 1.72569633e-01
- 1.84723666e-01 1.97195156e-01 2.09971545e-01 2.23039968e-01
- 2.36387266e-01 2.50000000e-01 2.63864463e-01 2.77966694e-01
- 2.92292493e-01 3.06827437e-01 3.21556889e-01 3.36466018e-01
- 3.51539812e-01 3.66763093e-01 3.82120532e-01 3.97596666e-01
- 4.13175911e-01 4.28842581e-01 4.44580900e-01 4.60375022e-01
- 4.76209042e-01 4.92067018e-01 5.07932982e-01 5.23790958e-01
- 5.39624978e-01 5.55419100e-01 5.71157419e-01 5.86824089e-01
- 6.02403334e-01 6.17879468e-01 6.33236907e-01 6.48460188e-01
- 6.63533982e-01 6.78443111e-01 6.93172563e-01 7.07707507e-01
- 7.22033306e-01 7.36135537e-01 7.50000000e-01 7.63612734e-01
- 7.76960032e-01 7.90028455e-01 8.02804844e-01 8.15276334e-01
- 8.27430367e-01 8.39254706e-01 8.50737444e-01 8.61867019e-01
- 8.72632225e-01 8.83022222e-01 8.93026547e-01 9.02635129e-01
- 9.11838291e-01 9.20626766e-01 9.28991707e-01 9.36924689e-01
- 9.44417724e-01 9.51463269e-01 9.58054229e-01 9.64183967e-01
- 9.69846310e-01 9.75035559e-01 9.79746487e-01 9.83974351e-01
- 9.87714893e-01 9.90964349e-01 9.93719444e-01 9.95977406e-01
- 9.97735961e-01 9.98993338e-01 9.99748271e-01 1.00000000e+00
- 9.99748271e-01 9.98993338e-01 9.97735961e-01 9.95977406e-01
- 9.93719444e-01 9.90964349e-01 9.87714893e-01 9.83974351e-01
- 9.79746487e-01 9.75035559e-01 9.69846310e-01 9.64183967e-01
- 9.58054229e-01 9.51463269e-01 9.44417724e-01 9.36924689e-01
- 9.28991707e-01 9.20626766e-01 9.11838291e-01 9.02635129e-01
- 8.93026547e-01 8.83022222e-01 8.72632225e-01 8.61867019e-01
- 8.50737444e-01 8.39254706e-01 8.27430367e-01 8.15276334e-01
- 8.02804844e-01 7.90028455e-01 7.76960032e-01 7.63612734e-01
- 7.50000000e-01 7.36135537e-01 7.22033306e-01 7.07707507e-01
- 6.93172563e-01 6.78443111e-01 6.63533982e-01 6.48460188e-01
- 6.33236907e-01 6.17879468e-01 6.02403334e-01 5.86824089e-01
- 5.71157419e-01 5.55419100e-01 5.39624978e-01 5.23790958e-01
- 5.07932982e-01 4.92067018e-01 4.76209042e-01 4.60375022e-01
- 4.44580900e-01 4.28842581e-01 4.13175911e-01 3.97596666e-01
- 3.82120532e-01 3.66763093e-01 3.51539812e-01 3.36466018e-01
- 3.21556889e-01 3.06827437e-01 2.92292493e-01 2.77966694e-01
- 2.63864463e-01 2.50000000e-01 2.36387266e-01 2.23039968e-01
- 2.09971545e-01 1.97195156e-01 1.84723666e-01 1.72569633e-01
- 1.60745294e-01 1.49262556e-01 1.38132981e-01 1.27367775e-01
- 1.16977778e-01 1.06973453e-01 9.73648712e-02 8.81617093e-02
- 7.93732336e-02 7.10082934e-02 6.30753115e-02 5.55822757e-02
- 4.85367309e-02 4.19457713e-02 3.58160335e-02 3.01536896e-02
- 2.49644411e-02 2.02535132e-02 1.60256493e-02 1.22851066e-02
- 9.03565137e-03 6.28055566e-03 4.02259358e-03 2.26403871e-03
- 1.00666176e-03 2.51728808e-04]  [ 0.          0.00744929  0.01357255  0.01926344  0.02487124  0.03075856
-  0.03681792  0.04299237  0.04930423  0.05567862  0.0621221   0.06865019
-  0.07515683  0.08177339  0.08843448  0.09502318  0.10158294  0.10804201
-  0.11442717  0.12068868  0.1268454   0.13286025  0.13874271  0.14442675
-  0.1499659   0.15526835  0.16037278  0.16525026  0.16983378  0.17414911
-  0.17817792  0.18184615  0.1852352   0.18835231  0.19120393  0.19377257
-  0.19611689  0.19825786  0.20020897  0.20196525  0.20351338  0.20486114
-  0.20601571  0.20699899  0.20780927  0.20852025  0.20915777  0.20966887
-  0.21003514  0.21026252  0.21036187  0.21034492  0.21021932  0.20998899
-  0.20966121  0.20924761  0.20875211  0.20818249  0.20755044  0.20686269
-  0.20612302  0.20533353  0.20450687  0.20364938  0.20274157  0.20177203
-  0.20076972  0.19977718  0.19878999  0.19780662  0.19683089  0.19587305
-  0.19493485  0.19402462  0.19314037  0.19228277  0.19145805  0.19066424
-  0.18989971  0.18916541  0.18846896  0.18780442  0.18716801  0.18656347
-  0.18597429  0.1853931   0.18484941  0.1843285   0.18383672  0.18338604
-  0.18297689  0.18258695  0.18224153  0.18194147  0.18168705  0.18147854
-  0.18131615  0.18120004  0.17720057  0.17585653  0.17586368  0.17588511
-  0.17592082  0.17597075  0.17603486  0.17611309  0.17620535  0.17631156
-  0.17642991  0.17618488  0.17591497  0.17562045  0.17517776  0.17425262
-  0.17326367  0.172009    0.17022882  0.16835171  0.16595716  0.16338862
-  0.1603589   0.15707335  0.15328911  0.14926076  0.14462456  0.13968022
-  0.13431823  0.12847746  0.12227945  0.11570177  0.10867115  0.10130358
-  0.09360951  0.08560391  0.07730179  0.06871083  0.05986824  0.0508001
-  0.04152731  0.03208553  0.02248215  0.01274621  0.00290732 -0.00700309
- -0.01694995 -0.02688723 -0.03677197 -0.04656643 -0.0562237  -0.06569077
- -0.0749222  -0.08385697 -0.09243735 -0.10055218 -0.10819397 -0.11527575
- -0.12173835 -0.12751264 -0.13246569 -0.13666148 -0.14005964 -0.1426379
- -0.14435696 -0.14524509 -0.14542002 -0.14491425 -0.14374524 -0.14203328
- -0.13988855 -0.13729713 -0.13433344 -0.13106364 -0.12744776 -0.12357836
- -0.11943916 -0.11508113 -0.11053219 -0.10580628 -0.10092838 -0.09594498
- -0.09081978 -0.08564486 -0.0804065  -0.07513051 -0.06986631 -0.06462719
- -0.05951496 -0.05446101 -0.04942663 -0.0443767  -0.03934783 -0.03433042
- -0.02923685 -0.02409943 -0.0190112  -0.01407878 -0.00902288 -0.00385672] [3440000000.0, 41800000000.0, 27700000000.0, 13600000000.0, 200000000000.0, 205000000000.0, 118000000000.0, 98460000.0, 114500000000.0, 1000000.0, 4560000.0] [3440000000.0, 14000000000.0, 13650000000.0, 13300000000.0, 200000000000.0, 205000000000.0, 118000000000.0, 98460000.0, 8390000000.0, 1000000.0, 4560000.0] [1323000000.0, 2630000000.0, 7200000000.0, 11800000000.0, 79300000000.0, 80000000000.0, 47600000000.0, 98460000.0, 5990000000.0, 384615.3846153846, 1520000.0] [0.3, 0.28, 0.39, 0.49, 0.3, 0.3, 0.3, 0.3, 0.27, 0.3, 0.49] [1235.0, 1920.0, 1850.0, 1780.0, 7800.0, 7850.0, 7200.0, 200.0, 1220.0, 1150.0, 1100.0] [0.         0.3835638  0.53111237 0.90085476 1.        ] [4. 4. 4. 4.] [1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1.] [5.00e-05 2.82e-03 2.00e-02 2.82e-03 5.00e-05 2.82e-03 4.23e-02 2.82e-03
- 5.00e-05 2.82e-03 9.00e-02 2.82e-03 5.00e-05 2.82e-03 7.05e-03 2.82e-03] [0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0.] [1 3 8 3 1 3 9 3 1 3 8 3 1 3 2 3] [0.         0.26428393 0.43085797 0.89876162 1.        ] [4. 4. 4. 4.] [1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1.] [5.00e-05 2.82e-03 2.00e-02 2.82e-03 5.00e-05 2.82e-03 4.23e-02 2.82e-03
- 5.00e-05 2.82e-03 9.00e-02 2.82e-03 5.00e-05 2.82e-03 7.05e-03 2.82e-03] [0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0.] [1 3 8 3 1 3 9 3 1 3 8 3 1 3 2 3] 2 [0.39872605 0.51756731] [3. 3.] [1. 1. 1. 1. 1. 1.] [0.002 0.04  0.002 0.002 0.04  0.002] [0. 0. 0. 0. 0. 0.] [4 8 4 4 8 4]
-
-
-
-9 4.426741607305162 0.0 0.0 0.4 [0.00000000e+00 2.51728808e-04 1.00666176e-03 2.26403871e-03
- 4.02259358e-03 6.28055566e-03 9.03565137e-03 1.22851066e-02
- 1.60256493e-02 2.02535132e-02 2.49644411e-02 3.01536896e-02
- 3.58160335e-02 4.19457713e-02 4.85367309e-02 5.55822757e-02
- 6.30753115e-02 7.10082934e-02 7.93732336e-02 8.81617093e-02
- 9.73648712e-02 1.06973453e-01 1.16977778e-01 1.27367775e-01
- 1.38132981e-01 1.49262556e-01 1.60745294e-01 1.72569633e-01
- 1.84723666e-01 1.97195156e-01 2.09971545e-01 2.23039968e-01
- 2.36387266e-01 2.50000000e-01 2.63864463e-01 2.77966694e-01
- 2.92292493e-01 3.06827437e-01 3.21556889e-01 3.36466018e-01
- 3.51539812e-01 3.66763093e-01 3.82120532e-01 3.97596666e-01
- 4.13175911e-01 4.28842581e-01 4.44580900e-01 4.60375022e-01
- 4.76209042e-01 4.92067018e-01 5.07932982e-01 5.23790958e-01
- 5.39624978e-01 5.55419100e-01 5.71157419e-01 5.86824089e-01
- 6.02403334e-01 6.17879468e-01 6.33236907e-01 6.48460188e-01
- 6.63533982e-01 6.78443111e-01 6.93172563e-01 7.07707507e-01
- 7.22033306e-01 7.36135537e-01 7.50000000e-01 7.63612734e-01
- 7.76960032e-01 7.90028455e-01 8.02804844e-01 8.15276334e-01
- 8.27430367e-01 8.39254706e-01 8.50737444e-01 8.61867019e-01
- 8.72632225e-01 8.83022222e-01 8.93026547e-01 9.02635129e-01
- 9.11838291e-01 9.20626766e-01 9.28991707e-01 9.36924689e-01
- 9.44417724e-01 9.51463269e-01 9.58054229e-01 9.64183967e-01
- 9.69846310e-01 9.75035559e-01 9.79746487e-01 9.83974351e-01
- 9.87714893e-01 9.90964349e-01 9.93719444e-01 9.95977406e-01
- 9.97735961e-01 9.98993338e-01 9.99748271e-01 1.00000000e+00
- 9.99748271e-01 9.98993338e-01 9.97735961e-01 9.95977406e-01
- 9.93719444e-01 9.90964349e-01 9.87714893e-01 9.83974351e-01
- 9.79746487e-01 9.75035559e-01 9.69846310e-01 9.64183967e-01
- 9.58054229e-01 9.51463269e-01 9.44417724e-01 9.36924689e-01
- 9.28991707e-01 9.20626766e-01 9.11838291e-01 9.02635129e-01
- 8.93026547e-01 8.83022222e-01 8.72632225e-01 8.61867019e-01
- 8.50737444e-01 8.39254706e-01 8.27430367e-01 8.15276334e-01
- 8.02804844e-01 7.90028455e-01 7.76960032e-01 7.63612734e-01
- 7.50000000e-01 7.36135537e-01 7.22033306e-01 7.07707507e-01
- 6.93172563e-01 6.78443111e-01 6.63533982e-01 6.48460188e-01
- 6.33236907e-01 6.17879468e-01 6.02403334e-01 5.86824089e-01
- 5.71157419e-01 5.55419100e-01 5.39624978e-01 5.23790958e-01
- 5.07932982e-01 4.92067018e-01 4.76209042e-01 4.60375022e-01
- 4.44580900e-01 4.28842581e-01 4.13175911e-01 3.97596666e-01
- 3.82120532e-01 3.66763093e-01 3.51539812e-01 3.36466018e-01
- 3.21556889e-01 3.06827437e-01 2.92292493e-01 2.77966694e-01
- 2.63864463e-01 2.50000000e-01 2.36387266e-01 2.23039968e-01
- 2.09971545e-01 1.97195156e-01 1.84723666e-01 1.72569633e-01
- 1.60745294e-01 1.49262556e-01 1.38132981e-01 1.27367775e-01
- 1.16977778e-01 1.06973453e-01 9.73648712e-02 8.81617093e-02
- 7.93732336e-02 7.10082934e-02 6.30753115e-02 5.55822757e-02
- 4.85367309e-02 4.19457713e-02 3.58160335e-02 3.01536896e-02
- 2.49644411e-02 2.02535132e-02 1.60256493e-02 1.22851066e-02
- 9.03565137e-03 6.28055566e-03 4.02259358e-03 2.26403871e-03
- 1.00666176e-03 2.51728808e-04]  [ 0.          0.00582524  0.01117969  0.01635185  0.0216645   0.02720048
-  0.0328703   0.03872104  0.04471667  0.05080825  0.05696983  0.06321645
-  0.06951343  0.07587644  0.0822741   0.0886603   0.09497965  0.10126314
-  0.10746353  0.11357749  0.11959595  0.12548425  0.13124935  0.1368519
-  0.14229093  0.1475579   0.15259218  0.15743452  0.16200497  0.16630259
-  0.1703259   0.17404611  0.17743651  0.18054031  0.18336988  0.18593886
-  0.18823634  0.19031756  0.19218504  0.19383784  0.19528492  0.19652585
-  0.19756939  0.19841998  0.1991337   0.19985581  0.20048302  0.20096194
-  0.20129829  0.20150053  0.20157667  0.20153561  0.20138432  0.20112899
-  0.20077784  0.20034047  0.19981935  0.19921608  0.19854862  0.19782185
-  0.19704451  0.19622415  0.19536674  0.1944689   0.19348498  0.19246622
-  0.19143445  0.19040663  0.18937532  0.18835293  0.18734168  0.18634082
-  0.18536189  0.18440597  0.18347346  0.18256788  0.18169251  0.18084709
-  0.18002988  0.17924528  0.17849506  0.17777859  0.17708814  0.17643334
-  0.17578542  0.17515495  0.17456516  0.17399554  0.17346148  0.17297204
-  0.17252592  0.17210148  0.17172597  0.17139975  0.17112317  0.17089649
-  0.17071995  0.17059372  0.16683397  0.16540276  0.16541248  0.16544166
-  0.16549024  0.1655582   0.16564545  0.16575191  0.16587747  0.16602202
-  0.16618539  0.16604265  0.16584611  0.16563166  0.16536303  0.16452648
-  0.16363223  0.16261215  0.16100205  0.15930427  0.15721686  0.154855
-  0.15217287  0.14911615  0.14571159  0.14193533  0.13770658  0.13318619
-  0.12809801  0.12269529  0.11692134  0.11065233  0.10405111  0.09711009
-  0.08981186  0.08217356  0.07425503  0.06607109  0.05764039  0.0489838
-  0.04012438  0.03108392  0.02189787  0.0125731   0.0031542  -0.00633534
- -0.01585552 -0.02537386 -0.03485134 -0.04425121 -0.0535357  -0.06264736
- -0.07153705 -0.08015239 -0.08844492 -0.0963176  -0.10370604 -0.11057916
- -0.11686319 -0.12250755 -0.12744325 -0.13155399 -0.13490781 -0.13748256
- -0.13927272 -0.14026549 -0.14047392 -0.14003736 -0.13900468 -0.13739089
- -0.13528895 -0.13280305 -0.12991733 -0.12669947 -0.12320357 -0.11939066
- -0.11536911 -0.11109376 -0.10666565 -0.10204956 -0.09731106 -0.09244682
- -0.08749216 -0.08247795 -0.07739433 -0.07229751 -0.06721367 -0.06214611
- -0.05713355 -0.05223165 -0.04738612 -0.04256354 -0.0377698  -0.03300351
- -0.02825564 -0.02344754 -0.01861896 -0.01383994 -0.00900147 -0.00402891] [3440000000.0, 41800000000.0, 27700000000.0, 13600000000.0, 200000000000.0, 205000000000.0, 118000000000.0, 98460000.0, 114500000000.0, 1000000.0, 4560000.0] [3440000000.0, 14000000000.0, 13650000000.0, 13300000000.0, 200000000000.0, 205000000000.0, 118000000000.0, 98460000.0, 8390000000.0, 1000000.0, 4560000.0] [1323000000.0, 2630000000.0, 7200000000.0, 11800000000.0, 79300000000.0, 80000000000.0, 47600000000.0, 98460000.0, 5990000000.0, 384615.3846153846, 1520000.0] [0.3, 0.28, 0.39, 0.49, 0.3, 0.3, 0.3, 0.3, 0.27, 0.3, 0.49] [1235.0, 1920.0, 1850.0, 1780.0, 7800.0, 7850.0, 7200.0, 200.0, 1220.0, 1150.0, 1100.0] [0.         0.37530761 0.52822711 0.8981405  1.        ] [4. 4. 4. 4.] [1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1.] [5.00e-05 2.82e-03 2.00e-02 2.82e-03 5.00e-05 2.82e-03 4.23e-02 2.82e-03
- 5.00e-05 2.82e-03 9.00e-02 2.82e-03 5.00e-05 2.82e-03 7.05e-03 2.82e-03] [0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0.] [1 3 8 3 1 3 9 3 1 3 8 3 1 3 2 3] [0.         0.26694576 0.43783019 0.896407   1.        ] [4. 4. 4. 4.] [1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1.] [5.00e-05 2.82e-03 2.00e-02 2.82e-03 5.00e-05 2.82e-03 4.23e-02 2.82e-03
- 5.00e-05 2.82e-03 9.00e-02 2.82e-03 5.00e-05 2.82e-03 7.05e-03 2.82e-03] [0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0.] [1 3 8 3 1 3 9 3 1 3 8 3 1 3 2 3] 2 [0.39101229 0.51420413] [3. 3.] [1. 1. 1. 1. 1. 1.] [0.002 0.04  0.002 0.002 0.04  0.002] [0. 0. 0. 0. 0. 0.] [4 8 4 4 8 4]
-
-25 2.416417405464169 0.0 0.0 0.4 [0.00000000e+00 2.51728808e-04 1.00666176e-03 2.26403871e-03
- 4.02259358e-03 6.28055566e-03 9.03565137e-03 1.22851066e-02
- 1.60256493e-02 2.02535132e-02 2.49644411e-02 3.01536896e-02
- 3.58160335e-02 4.19457713e-02 4.85367309e-02 5.55822757e-02
- 6.30753115e-02 7.10082934e-02 7.93732336e-02 8.81617093e-02
- 9.73648712e-02 1.06973453e-01 1.16977778e-01 1.27367775e-01
- 1.38132981e-01 1.49262556e-01 1.60745294e-01 1.72569633e-01
- 1.84723666e-01 1.97195156e-01 2.09971545e-01 2.23039968e-01
- 2.36387266e-01 2.50000000e-01 2.63864463e-01 2.77966694e-01
- 2.92292493e-01 3.06827437e-01 3.21556889e-01 3.36466018e-01
- 3.51539812e-01 3.66763093e-01 3.82120532e-01 3.97596666e-01
- 4.13175911e-01 4.28842581e-01 4.44580900e-01 4.60375022e-01
- 4.76209042e-01 4.92067018e-01 5.07932982e-01 5.23790958e-01
- 5.39624978e-01 5.55419100e-01 5.71157419e-01 5.86824089e-01
- 6.02403334e-01 6.17879468e-01 6.33236907e-01 6.48460188e-01
- 6.63533982e-01 6.78443111e-01 6.93172563e-01 7.07707507e-01
- 7.22033306e-01 7.36135537e-01 7.50000000e-01 7.63612734e-01
- 7.76960032e-01 7.90028455e-01 8.02804844e-01 8.15276334e-01
- 8.27430367e-01 8.39254706e-01 8.50737444e-01 8.61867019e-01
- 8.72632225e-01 8.83022222e-01 8.93026547e-01 9.02635129e-01
- 9.11838291e-01 9.20626766e-01 9.28991707e-01 9.36924689e-01
- 9.44417724e-01 9.51463269e-01 9.58054229e-01 9.64183967e-01
- 9.69846310e-01 9.75035559e-01 9.79746487e-01 9.83974351e-01
- 9.87714893e-01 9.90964349e-01 9.93719444e-01 9.95977406e-01
- 9.97735961e-01 9.98993338e-01 9.99748271e-01 1.00000000e+00
- 9.99748271e-01 9.98993338e-01 9.97735961e-01 9.95977406e-01
- 9.93719444e-01 9.90964349e-01 9.87714893e-01 9.83974351e-01
- 9.79746487e-01 9.75035559e-01 9.69846310e-01 9.64183967e-01
- 9.58054229e-01 9.51463269e-01 9.44417724e-01 9.36924689e-01
- 9.28991707e-01 9.20626766e-01 9.11838291e-01 9.02635129e-01
- 8.93026547e-01 8.83022222e-01 8.72632225e-01 8.61867019e-01
- 8.50737444e-01 8.39254706e-01 8.27430367e-01 8.15276334e-01
- 8.02804844e-01 7.90028455e-01 7.76960032e-01 7.63612734e-01
- 7.50000000e-01 7.36135537e-01 7.22033306e-01 7.07707507e-01
- 6.93172563e-01 6.78443111e-01 6.63533982e-01 6.48460188e-01
- 6.33236907e-01 6.17879468e-01 6.02403334e-01 5.86824089e-01
- 5.71157419e-01 5.55419100e-01 5.39624978e-01 5.23790958e-01
- 5.07932982e-01 4.92067018e-01 4.76209042e-01 4.60375022e-01
- 4.44580900e-01 4.28842581e-01 4.13175911e-01 3.97596666e-01
- 3.82120532e-01 3.66763093e-01 3.51539812e-01 3.36466018e-01
- 3.21556889e-01 3.06827437e-01 2.92292493e-01 2.77966694e-01
- 2.63864463e-01 2.50000000e-01 2.36387266e-01 2.23039968e-01
- 2.09971545e-01 1.97195156e-01 1.84723666e-01 1.72569633e-01
- 1.60745294e-01 1.49262556e-01 1.38132981e-01 1.27367775e-01
- 1.16977778e-01 1.06973453e-01 9.73648712e-02 8.81617093e-02
- 7.93732336e-02 7.10082934e-02 6.30753115e-02 5.55822757e-02
- 4.85367309e-02 4.19457713e-02 3.58160335e-02 3.01536896e-02
- 2.49644411e-02 2.02535132e-02 1.60256493e-02 1.22851066e-02
- 9.03565137e-03 6.28055566e-03 4.02259358e-03 2.26403871e-03
- 1.00666176e-03 2.51728808e-04]  [ 0.00000000e+00  3.89663011e-03  7.15832472e-03  1.04071687e-02
-  1.37187070e-02  1.70169235e-02  2.03816954e-02  2.38253168e-02
-  2.73429567e-02  3.09256169e-02  3.45697171e-02  3.82581262e-02
-  4.19893097e-02  4.57524453e-02  4.95345786e-02  5.33262749e-02
-  5.71055475e-02  6.08747321e-02  6.46289274e-02  6.83614045e-02
-  7.20393685e-02  7.56837248e-02  7.92705878e-02  8.27925740e-02
-  8.62418977e-02  8.96145290e-02  9.28895272e-02  9.60801274e-02
-  9.91494599e-02  1.02109037e-01  1.04956070e-01  1.07670927e-01
-  1.10233124e-01  1.12654041e-01  1.14923978e-01  1.17032274e-01
-  1.18970821e-01  1.20725475e-01  1.22276281e-01  1.23620052e-01
-  1.24742841e-01  1.25628619e-01  1.26242165e-01  1.26541857e-01
-  1.26502985e-01  1.26138748e-01  1.25506815e-01  1.24629105e-01
-  1.23543638e-01  1.22251148e-01  1.20760838e-01  1.19091085e-01
-  1.17260227e-01  1.15244859e-01  1.13072291e-01  1.10771459e-01
-  1.08356618e-01  1.05836808e-01  1.03177657e-01  1.00444724e-01
-  9.76512404e-02  9.47849469e-02  9.18579990e-02  8.89102809e-02
-  8.59313763e-02  8.29193482e-02  7.99196022e-02  7.69093901e-02
-  7.39184628e-02  7.09618552e-02  6.80264265e-02  6.51517428e-02
-  6.23171743e-02  5.95558829e-02  5.68574425e-02  5.42408937e-02
-  5.17039177e-02  4.92553651e-02  4.68996369e-02  4.46371545e-02
-  4.24748428e-02  4.04133280e-02  3.84511620e-02  3.65288035e-02
-  3.47053979e-02  3.29908878e-02  3.13578501e-02  2.98283334e-02
-  2.84154428e-02  2.71206008e-02  2.59222666e-02  2.48346973e-02
-  2.38724857e-02  2.30366007e-02  2.23278840e-02  2.17470493e-02
-  2.12946813e-02  2.09712355e-02  2.07770377e-02  2.07122834e-02
-  2.07353506e-02  2.08045290e-02  2.09197489e-02  2.10808944e-02
-  2.12878031e-02  2.15402668e-02  2.18380311e-02  2.21807964e-02
-  2.25682174e-02  2.29299891e-02  2.31513860e-02  2.33929674e-02
-  2.36544898e-02  2.38798136e-02  2.38742478e-02  2.38683285e-02
-  2.38577280e-02  2.34070523e-02  2.29335579e-02  2.23572265e-02
-  2.15703772e-02  2.07511203e-02  1.96991947e-02  1.85886033e-02
-  1.72902472e-02  1.58924409e-02  1.43286675e-02  1.26676025e-02
-  1.08407040e-02  8.93732306e-03  6.86647240e-03  4.71944400e-03
-  2.45993413e-03  1.03472168e-04 -2.32577079e-03 -4.84403323e-03
- -7.42429283e-03 -1.00585119e-02 -1.27448650e-02 -1.54664919e-02
- -1.82113963e-02 -2.09697384e-02 -2.37165559e-02 -2.64464729e-02
- -2.91489287e-02 -3.18109278e-02 -3.44192394e-02 -3.69496809e-02
- -3.93931035e-02 -4.17368877e-02 -4.39641669e-02 -4.60656602e-02
- -4.80222635e-02 -4.98027887e-02 -5.13701967e-02 -5.26755439e-02
- -5.37034983e-02 -5.45002885e-02 -5.51111934e-02 -5.55556694e-02
- -5.58516788e-02 -5.60111926e-02 -5.60422013e-02 -5.59508581e-02
- -5.57420524e-02 -5.54182122e-02 -5.49928397e-02 -5.44721106e-02
- -5.38616705e-02 -5.31474608e-02 -5.23525176e-02 -5.14809920e-02
- -5.05189459e-02 -4.94902304e-02 -4.83820848e-02 -4.72089241e-02
- -4.59614560e-02 -4.46564420e-02 -4.32833385e-02 -4.18532043e-02
- -4.03610201e-02 -3.87963109e-02 -3.71754477e-02 -3.54982111e-02
- -3.37646763e-02 -3.19742986e-02 -3.01256781e-02 -2.82093459e-02
- -2.62189353e-02 -2.41542438e-02 -2.19991649e-02 -1.97453398e-02
- -1.73810180e-02 -1.48844842e-02 -1.22026982e-02 -9.20670200e-03
- -6.02015625e-03 -2.75235211e-03] [3440000000.0, 41800000000.0, 27700000000.0, 13600000000.0, 200000000000.0, 205000000000.0, 118000000000.0, 98460000.0, 114500000000.0, 1000000.0, 4560000.0] [3440000000.0, 14000000000.0, 13650000000.0, 13300000000.0, 200000000000.0, 205000000000.0, 118000000000.0, 98460000.0, 8390000000.0, 1000000.0, 4560000.0] [1323000000.0, 2630000000.0, 7200000000.0, 11800000000.0, 79300000000.0, 80000000000.0, 47600000000.0, 98460000.0, 5990000000.0, 384615.3846153846, 1520000.0] [0.3, 0.28, 0.39, 0.49, 0.3, 0.3, 0.3, 0.3, 0.27, 0.3, 0.49] [1235.0, 1920.0, 1850.0, 1780.0, 7800.0, 7850.0, 7200.0, 200.0, 1220.0, 1150.0, 1100.0] [0.         0.24737313 0.556296   0.86769712 1.        ] [4. 4. 4. 4.] [1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1.] [5.00000000e-05 2.82000000e-03 2.00000000e-02 2.82000000e-03
- 5.00000000e-05 2.82000000e-03 1.53951732e-02 2.82000000e-03
- 5.00000000e-05 2.82000000e-03 2.00000000e-02 2.82000000e-03
- 5.00000000e-05 2.82000000e-03 1.39341227e-04 2.82000000e-03] [0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0.] [1 3 8 3 1 3 9 3 1 3 8 3 1 3 2 3] [0.         0.24084585 0.54987261 0.87115473 1.        ] [4. 4. 4. 4.] [1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1.] [5.00000000e-05 2.82000000e-03 2.00000000e-02 2.82000000e-03
- 5.00000000e-05 2.82000000e-03 1.53951732e-02 2.82000000e-03
- 5.00000000e-05 2.82000000e-03 2.00000000e-02 2.82000000e-03
- 5.00000000e-05 2.82000000e-03 1.39341227e-04 2.82000000e-03] [0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0.] [1 3 8 3 1 3 9 3 1 3 8 3 1 3 2 3] 2 [0.27836362 0.52613213] [3. 3.] [1. 1. 1. 1. 1. 1.] [0.002 0.04  0.002 0.002 0.04  0.002] [0. 0. 0. 0. 0. 0.] [4 8 4 4 8 4]
-
-26 2.280796764281841 0.0 0.0 0.4 [0.00000000e+00 2.51728808e-04 1.00666176e-03 2.26403871e-03
- 4.02259358e-03 6.28055566e-03 9.03565137e-03 1.22851066e-02
- 1.60256493e-02 2.02535132e-02 2.49644411e-02 3.01536896e-02
- 3.58160335e-02 4.19457713e-02 4.85367309e-02 5.55822757e-02
- 6.30753115e-02 7.10082934e-02 7.93732336e-02 8.81617093e-02
- 9.73648712e-02 1.06973453e-01 1.16977778e-01 1.27367775e-01
- 1.38132981e-01 1.49262556e-01 1.60745294e-01 1.72569633e-01
- 1.84723666e-01 1.97195156e-01 2.09971545e-01 2.23039968e-01
- 2.36387266e-01 2.50000000e-01 2.63864463e-01 2.77966694e-01
- 2.92292493e-01 3.06827437e-01 3.21556889e-01 3.36466018e-01
- 3.51539812e-01 3.66763093e-01 3.82120532e-01 3.97596666e-01
- 4.13175911e-01 4.28842581e-01 4.44580900e-01 4.60375022e-01
- 4.76209042e-01 4.92067018e-01 5.07932982e-01 5.23790958e-01
- 5.39624978e-01 5.55419100e-01 5.71157419e-01 5.86824089e-01
- 6.02403334e-01 6.17879468e-01 6.33236907e-01 6.48460188e-01
- 6.63533982e-01 6.78443111e-01 6.93172563e-01 7.07707507e-01
- 7.22033306e-01 7.36135537e-01 7.50000000e-01 7.63612734e-01
- 7.76960032e-01 7.90028455e-01 8.02804844e-01 8.15276334e-01
- 8.27430367e-01 8.39254706e-01 8.50737444e-01 8.61867019e-01
- 8.72632225e-01 8.83022222e-01 8.93026547e-01 9.02635129e-01
- 9.11838291e-01 9.20626766e-01 9.28991707e-01 9.36924689e-01
- 9.44417724e-01 9.51463269e-01 9.58054229e-01 9.64183967e-01
- 9.69846310e-01 9.75035559e-01 9.79746487e-01 9.83974351e-01
- 9.87714893e-01 9.90964349e-01 9.93719444e-01 9.95977406e-01
- 9.97735961e-01 9.98993338e-01 9.99748271e-01 1.00000000e+00
- 9.99748271e-01 9.98993338e-01 9.97735961e-01 9.95977406e-01
- 9.93719444e-01 9.90964349e-01 9.87714893e-01 9.83974351e-01
- 9.79746487e-01 9.75035559e-01 9.69846310e-01 9.64183967e-01
- 9.58054229e-01 9.51463269e-01 9.44417724e-01 9.36924689e-01
- 9.28991707e-01 9.20626766e-01 9.11838291e-01 9.02635129e-01
- 8.93026547e-01 8.83022222e-01 8.72632225e-01 8.61867019e-01
- 8.50737444e-01 8.39254706e-01 8.27430367e-01 8.15276334e-01
- 8.02804844e-01 7.90028455e-01 7.76960032e-01 7.63612734e-01
- 7.50000000e-01 7.36135537e-01 7.22033306e-01 7.07707507e-01
- 6.93172563e-01 6.78443111e-01 6.63533982e-01 6.48460188e-01
- 6.33236907e-01 6.17879468e-01 6.02403334e-01 5.86824089e-01
- 5.71157419e-01 5.55419100e-01 5.39624978e-01 5.23790958e-01
- 5.07932982e-01 4.92067018e-01 4.76209042e-01 4.60375022e-01
- 4.44580900e-01 4.28842581e-01 4.13175911e-01 3.97596666e-01
- 3.82120532e-01 3.66763093e-01 3.51539812e-01 3.36466018e-01
- 3.21556889e-01 3.06827437e-01 2.92292493e-01 2.77966694e-01
- 2.63864463e-01 2.50000000e-01 2.36387266e-01 2.23039968e-01
- 2.09971545e-01 1.97195156e-01 1.84723666e-01 1.72569633e-01
- 1.60745294e-01 1.49262556e-01 1.38132981e-01 1.27367775e-01
- 1.16977778e-01 1.06973453e-01 9.73648712e-02 8.81617093e-02
- 7.93732336e-02 7.10082934e-02 6.30753115e-02 5.55822757e-02
- 4.85367309e-02 4.19457713e-02 3.58160335e-02 3.01536896e-02
- 2.49644411e-02 2.02535132e-02 1.60256493e-02 1.22851066e-02
- 9.03565137e-03 6.28055566e-03 4.02259358e-03 2.26403871e-03
- 1.00666176e-03 2.51728808e-04] [ 0.          0.00372158  0.00700204  0.01023018  0.01352453  0.01681594
-  0.02015436  0.02356812  0.02705336  0.03060156  0.03421289  0.03785971
-  0.04154652  0.045263    0.04899596  0.05272975  0.05645545  0.06016859
-  0.06386435  0.06752981  0.07114545  0.07472495  0.07823864  0.08169244
-  0.08506463  0.08836552  0.09155996  0.09466814  0.09766131  0.10053548
-  0.10329552  0.10593084  0.10840373  0.11073402  0.1129124   0.11492809
-  0.11677305  0.11843365  0.11989775  0.1211508   0.12217907  0.1229714
-  0.12349327  0.12370224  0.1235612   0.12309462  0.12236525  0.12138878
-  0.12020168  0.11880784  0.1172149   0.11544107  0.11350571  0.11137602
-  0.1090984   0.10669198  0.10417107  0.1015381   0.09877332  0.09593468
-  0.0930355   0.09005905  0.08702766  0.08397617  0.08089077  0.07777721
-  0.07467688  0.07156491  0.06847627  0.06542176  0.06239203  0.05942475
-  0.05649885  0.05364923  0.05086443  0.04816435  0.04554639  0.04301967
-  0.04058864  0.03825388  0.0360222   0.0338946   0.03186954  0.0298885
-  0.02800853  0.02624082  0.02455849  0.02298256  0.02152679  0.02019266
-  0.01895893  0.01783923  0.01684859  0.01598801  0.01525835  0.01466036
-  0.01419462  0.01386162  0.01366169  0.01359502  0.01361989  0.01369449
-  0.01381874  0.01399251  0.01421563  0.01448787  0.01480897  0.01517859
-  0.01559637  0.01599097  0.01624937  0.01653132  0.01683655  0.01710757
-  0.01715213  0.01719953  0.01724363  0.01685263  0.01644182  0.01593045
-  0.01521238  0.01446475  0.01348709  0.01245385  0.01123564  0.00992068
-  0.00844263  0.00686957  0.00513361  0.00332356  0.00134792 -0.00070123
- -0.00286133 -0.00511585 -0.00744142 -0.00985473 -0.01232804 -0.01485397
- -0.01743094 -0.02004186 -0.02267527 -0.0253215  -0.02795487 -0.03057145
- -0.03316039 -0.03570885 -0.0382037  -0.04061873 -0.04294896 -0.04518025
- -0.04729581 -0.04928612 -0.05113306 -0.05280544 -0.05426708 -0.05546874
- -0.05639288 -0.05708535 -0.05759244 -0.05793383 -0.05812743 -0.05818583
- -0.05811678 -0.05792686 -0.0576237  -0.05720403 -0.05668455 -0.05607135
- -0.0553714  -0.05457013 -0.05368855 -0.05273499 -0.05169066 -0.05058193
- -0.04939862 -0.04815029 -0.04683267 -0.04545852 -0.04401953 -0.0425276
- -0.04097449 -0.03935307 -0.0376791  -0.03595197 -0.03417185 -0.03233815
- -0.03044721 -0.02849298 -0.02646998 -0.0243758  -0.02219625 -0.01992195
- -0.01754124 -0.01503269 -0.01234479 -0.00936476 -0.0061621  -0.00288694] [3440000000.0, 41800000000.0, 27700000000.0, 13600000000.0, 200000000000.0, 205000000000.0, 118000000000.0, 98460000.0, 114500000000.0, 1000000.0, 4560000.0] [3440000000.0, 14000000000.0, 13650000000.0, 13300000000.0, 200000000000.0, 205000000000.0, 118000000000.0, 98460000.0, 8390000000.0, 1000000.0, 4560000.0] [1323000000.0, 2630000000.0, 7200000000.0, 11800000000.0, 79300000000.0, 80000000000.0, 47600000000.0, 98460000.0, 5990000000.0, 384615.3846153846, 1520000.0] [0.3, 0.28, 0.39, 0.49, 0.3, 0.3, 0.3, 0.3, 0.27, 0.3, 0.49] [1235.0, 1920.0, 1850.0, 1780.0, 7800.0, 7850.0, 7200.0, 200.0, 1220.0, 1150.0, 1100.0] [0.         0.23728968 0.56467374 0.88762136 1.        ] [4. 4. 4. 4.] [1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1.] [5.00000000e-05 2.82000000e-03 2.00000000e-02 2.82000000e-03
- 5.00000000e-05 2.82000000e-03 1.19246502e-02 2.82000000e-03
- 5.00000000e-05 2.82000000e-03 2.00000000e-02 2.82000000e-03
- 5.00000000e-05 2.82000000e-03 8.06287937e-05 2.82000000e-03] [0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0.] [1 3 8 3 1 3 9 3 1 3 8 3 1 3 2 3] [0.         0.23334333 0.56059153 0.89082564 1.        ] [4. 4. 4. 4.] [1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1.] [5.00000000e-05 2.82000000e-03 2.00000000e-02 2.82000000e-03
- 5.00000000e-05 2.82000000e-03 1.19246502e-02 2.82000000e-03
- 5.00000000e-05 2.82000000e-03 2.00000000e-02 2.82000000e-03
- 5.00000000e-05 2.82000000e-03 8.06287937e-05 2.82000000e-03] [0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0.] [1 3 8 3 1 3 9 3 1 3 8 3 1 3 2 3] 2 [0.27011043 0.53279921] [3. 3.] [1. 1. 1. 1. 1. 1.] [0.002 0.04  0.002 0.002 0.04  0.002] [0. 0. 0. 0. 0. 0.] [4 8 4 4 8 4]
-
-
-27 2.1030278893097076 0.0 0.0 0.4 [0.00000000e+00 2.51728808e-04 1.00666176e-03 2.26403871e-03
- 4.02259358e-03 6.28055566e-03 9.03565137e-03 1.22851066e-02
- 1.60256493e-02 2.02535132e-02 2.49644411e-02 3.01536896e-02
- 3.58160335e-02 4.19457713e-02 4.85367309e-02 5.55822757e-02
- 6.30753115e-02 7.10082934e-02 7.93732336e-02 8.81617093e-02
- 9.73648712e-02 1.06973453e-01 1.16977778e-01 1.27367775e-01
- 1.38132981e-01 1.49262556e-01 1.60745294e-01 1.72569633e-01
- 1.84723666e-01 1.97195156e-01 2.09971545e-01 2.23039968e-01
- 2.36387266e-01 2.50000000e-01 2.63864463e-01 2.77966694e-01
- 2.92292493e-01 3.06827437e-01 3.21556889e-01 3.36466018e-01
- 3.51539812e-01 3.66763093e-01 3.82120532e-01 3.97596666e-01
- 4.13175911e-01 4.28842581e-01 4.44580900e-01 4.60375022e-01
- 4.76209042e-01 4.92067018e-01 5.07932982e-01 5.23790958e-01
- 5.39624978e-01 5.55419100e-01 5.71157419e-01 5.86824089e-01
- 6.02403334e-01 6.17879468e-01 6.33236907e-01 6.48460188e-01
- 6.63533982e-01 6.78443111e-01 6.93172563e-01 7.07707507e-01
- 7.22033306e-01 7.36135537e-01 7.50000000e-01 7.63612734e-01
- 7.76960032e-01 7.90028455e-01 8.02804844e-01 8.15276334e-01
- 8.27430367e-01 8.39254706e-01 8.50737444e-01 8.61867019e-01
- 8.72632225e-01 8.83022222e-01 8.93026547e-01 9.02635129e-01
- 9.11838291e-01 9.20626766e-01 9.28991707e-01 9.36924689e-01
- 9.44417724e-01 9.51463269e-01 9.58054229e-01 9.64183967e-01
- 9.69846310e-01 9.75035559e-01 9.79746487e-01 9.83974351e-01
- 9.87714893e-01 9.90964349e-01 9.93719444e-01 9.95977406e-01
- 9.97735961e-01 9.98993338e-01 9.99748271e-01 1.00000000e+00
- 9.99748271e-01 9.98993338e-01 9.97735961e-01 9.95977406e-01
- 9.93719444e-01 9.90964349e-01 9.87714893e-01 9.83974351e-01
- 9.79746487e-01 9.75035559e-01 9.69846310e-01 9.64183967e-01
- 9.58054229e-01 9.51463269e-01 9.44417724e-01 9.36924689e-01
- 9.28991707e-01 9.20626766e-01 9.11838291e-01 9.02635129e-01
- 8.93026547e-01 8.83022222e-01 8.72632225e-01 8.61867019e-01
- 8.50737444e-01 8.39254706e-01 8.27430367e-01 8.15276334e-01
- 8.02804844e-01 7.90028455e-01 7.76960032e-01 7.63612734e-01
- 7.50000000e-01 7.36135537e-01 7.22033306e-01 7.07707507e-01
- 6.93172563e-01 6.78443111e-01 6.63533982e-01 6.48460188e-01
- 6.33236907e-01 6.17879468e-01 6.02403334e-01 5.86824089e-01
- 5.71157419e-01 5.55419100e-01 5.39624978e-01 5.23790958e-01
- 5.07932982e-01 4.92067018e-01 4.76209042e-01 4.60375022e-01
- 4.44580900e-01 4.28842581e-01 4.13175911e-01 3.97596666e-01
- 3.82120532e-01 3.66763093e-01 3.51539812e-01 3.36466018e-01
- 3.21556889e-01 3.06827437e-01 2.92292493e-01 2.77966694e-01
- 2.63864463e-01 2.50000000e-01 2.36387266e-01 2.23039968e-01
- 2.09971545e-01 1.97195156e-01 1.84723666e-01 1.72569633e-01
- 1.60745294e-01 1.49262556e-01 1.38132981e-01 1.27367775e-01
- 1.16977778e-01 1.06973453e-01 9.73648712e-02 8.81617093e-02
- 7.93732336e-02 7.10082934e-02 6.30753115e-02 5.55822757e-02
- 4.85367309e-02 4.19457713e-02 3.58160335e-02 3.01536896e-02
- 2.49644411e-02 2.02535132e-02 1.60256493e-02 1.22851066e-02
- 9.03565137e-03 6.28055566e-03 4.02259358e-03 2.26403871e-03
- 1.00666176e-03 2.51728808e-04] [ 0.          0.00357062  0.00683313  0.01006896  0.01334657  0.01663043
-  0.01994454  0.02333035  0.02678525  0.03030045  0.03387648  0.03748973
-  0.04113502  0.04480785  0.04848986  0.05217441  0.05584994  0.05951057
-  0.06315159  0.06675428  0.0703115   0.07383019  0.07727531  0.08066515
-  0.083965    0.08719858  0.09031744  0.09334829  0.09626995  0.09906394
-  0.10174231  0.10429544  0.10669193  0.10893709  0.11102944  0.11295809
-  0.1147151   0.11628727  0.11766236  0.11883217  0.11977692  0.12048134
-  0.12091694  0.12103036  0.12079512  0.12024033  0.11941919  0.11834966
-  0.11706679  0.11557745  0.11388776  0.11201584  0.1099774   0.10774444
-  0.10536774  0.10286171  0.10024076  0.097501    0.09463656  0.09169815
-  0.08869923  0.0856188   0.08248881  0.07933936  0.0761534   0.072944
-  0.06974868  0.06654062  0.0633597   0.06021272  0.05709394  0.05403922
-  0.05102705  0.04809401  0.04522764  0.04244866  0.0397542   0.03715365
-  0.0346515   0.03224846  0.02995125  0.02776124  0.02567676  0.02364031
-  0.02170689  0.01988894  0.01816003  0.01654026  0.01504399  0.01367274
-  0.01240556  0.0112555   0.010238    0.00935409  0.00860465  0.00799044
-  0.00751208  0.00717005  0.00696469  0.00689622  0.00692279  0.0070025
-  0.00713526  0.00732093  0.00755934  0.00785023  0.00819331  0.00858825
-  0.00903464  0.00946018  0.00975344  0.01007343  0.01041983  0.01073389
-  0.01082565  0.01092324  0.01101884  0.01068397  0.01033214  0.00988187
-  0.0092285   0.00854821  0.00764038  0.00667987  0.00553699  0.00429989
-  0.0029024   0.00141204 -0.00023848 -0.00196084 -0.00384703 -0.00580424
- -0.00787096 -0.01002974 -0.01225793 -0.01457267 -0.01694554 -0.01936977
- -0.02184401 -0.02435097 -0.02687967 -0.0294206  -0.03194744 -0.03445755
- -0.03693989 -0.03938169 -0.04176996 -0.04407658 -0.04630045 -0.04842604
- -0.05043662 -0.05232245 -0.05406631 -0.05563702 -0.05699921 -0.05810336
- -0.05892984 -0.05952417 -0.05993357 -0.06017794 -0.06027516 -0.06023837
- -0.06007513 -0.05979239 -0.05939879 -0.05889046 -0.05828234 -0.05758204
- -0.05679676 -0.05591448 -0.05495126 -0.0539181  -0.05279851 -0.05161436
- -0.05036018 -0.04904099 -0.04765723 -0.04621705 -0.04471673 -0.04316651
- -0.04155523 -0.03988097 -0.03815688 -0.03638265 -0.03455854 -0.03268244
- -0.03075156 -0.02876296 -0.02670914 -0.0245871  -0.02238443 -0.02009082
- -0.0176947  -0.01517495 -0.01248166 -0.0095185  -0.00630196 -0.0030262 ] [3440000000.0, 41800000000.0, 27700000000.0, 13600000000.0, 200000000000.0, 205000000000.0, 118000000000.0, 98460000.0, 114500000000.0, 1000000.0, 4560000.0] [3440000000.0, 14000000000.0, 13650000000.0, 13300000000.0, 200000000000.0, 205000000000.0, 118000000000.0, 98460000.0, 8390000000.0, 1000000.0, 4560000.0] [1323000000.0, 2630000000.0, 7200000000.0, 11800000000.0, 79300000000.0, 80000000000.0, 47600000000.0, 98460000.0, 5990000000.0, 384615.3846153846, 1520000.0] [0.3, 0.28, 0.39, 0.49, 0.3, 0.3, 0.3, 0.3, 0.27, 0.3, 0.49] [1235.0, 1920.0, 1850.0, 1780.0, 7800.0, 7850.0, 7200.0, 200.0, 1220.0, 1150.0, 1100.0] [0.         0.22260296 0.57772209 0.9107544  1.        ] [4. 4. 4. 4.] [1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1.] [5.00000000e-05 2.82000000e-03 2.00000000e-02 2.82000000e-03
- 5.00000000e-05 2.82000000e-03 8.26683295e-03 2.82000000e-03
- 5.00000000e-05 2.82000000e-03 2.00000000e-02 2.82000000e-03
- 5.00000000e-05 2.82000000e-03 3.68347320e-05 2.82000000e-03] [0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0.] [1 3 8 3 1 3 9 3 1 3 8 3 1 3 2 3] [0.         0.22115142 0.57580238 0.91348983 1.        ] [4. 4. 4. 4.] [1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1.] [5.00000000e-05 2.82000000e-03 2.00000000e-02 2.82000000e-03
- 5.00000000e-05 2.82000000e-03 8.26683295e-03 2.82000000e-03
- 5.00000000e-05 2.82000000e-03 2.00000000e-02 2.82000000e-03
- 5.00000000e-05 2.82000000e-03 3.68347320e-05 2.82000000e-03] [0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0.] [1 3 8 3 1 3 9 3 1 3 8 3 1 3 2 3] 2 [0.25817165 0.54326523] [3. 3.] [1. 1. 1. 1. 1. 1.] [0.002 0.04  0.002 0.002 0.04  0.002] [0. 0. 0. 0. 0. 0.] [4 8 4 4 8 4]
-
-
-
-28 1.5245990923819477 0.0 0.0 0.4 [0.00000000e+00 2.51728808e-04 1.00666176e-03 2.26403871e-03
- 4.02259358e-03 6.28055566e-03 9.03565137e-03 1.22851066e-02
- 1.60256493e-02 2.02535132e-02 2.49644411e-02 3.01536896e-02
- 3.58160335e-02 4.19457713e-02 4.85367309e-02 5.55822757e-02
- 6.30753115e-02 7.10082934e-02 7.93732336e-02 8.81617093e-02
- 9.73648712e-02 1.06973453e-01 1.16977778e-01 1.27367775e-01
- 1.38132981e-01 1.49262556e-01 1.60745294e-01 1.72569633e-01
- 1.84723666e-01 1.97195156e-01 2.09971545e-01 2.23039968e-01
- 2.36387266e-01 2.50000000e-01 2.63864463e-01 2.77966694e-01
- 2.92292493e-01 3.06827437e-01 3.21556889e-01 3.36466018e-01
- 3.51539812e-01 3.66763093e-01 3.82120532e-01 3.97596666e-01
- 4.13175911e-01 4.28842581e-01 4.44580900e-01 4.60375022e-01
- 4.76209042e-01 4.92067018e-01 5.07932982e-01 5.23790958e-01
- 5.39624978e-01 5.55419100e-01 5.71157419e-01 5.86824089e-01
- 6.02403334e-01 6.17879468e-01 6.33236907e-01 6.48460188e-01
- 6.63533982e-01 6.78443111e-01 6.93172563e-01 7.07707507e-01
- 7.22033306e-01 7.36135537e-01 7.50000000e-01 7.63612734e-01
- 7.76960032e-01 7.90028455e-01 8.02804844e-01 8.15276334e-01
- 8.27430367e-01 8.39254706e-01 8.50737444e-01 8.61867019e-01
- 8.72632225e-01 8.83022222e-01 8.93026547e-01 9.02635129e-01
- 9.11838291e-01 9.20626766e-01 9.28991707e-01 9.36924689e-01
- 9.44417724e-01 9.51463269e-01 9.58054229e-01 9.64183967e-01
- 9.69846310e-01 9.75035559e-01 9.79746487e-01 9.83974351e-01
- 9.87714893e-01 9.90964349e-01 9.93719444e-01 9.95977406e-01
- 9.97735961e-01 9.98993338e-01 9.99748271e-01 1.00000000e+00
- 9.99748271e-01 9.98993338e-01 9.97735961e-01 9.95977406e-01
- 9.93719444e-01 9.90964349e-01 9.87714893e-01 9.83974351e-01
- 9.79746487e-01 9.75035559e-01 9.69846310e-01 9.64183967e-01
- 9.58054229e-01 9.51463269e-01 9.44417724e-01 9.36924689e-01
- 9.28991707e-01 9.20626766e-01 9.11838291e-01 9.02635129e-01
- 8.93026547e-01 8.83022222e-01 8.72632225e-01 8.61867019e-01
- 8.50737444e-01 8.39254706e-01 8.27430367e-01 8.15276334e-01
- 8.02804844e-01 7.90028455e-01 7.76960032e-01 7.63612734e-01
- 7.50000000e-01 7.36135537e-01 7.22033306e-01 7.07707507e-01
- 6.93172563e-01 6.78443111e-01 6.63533982e-01 6.48460188e-01
- 6.33236907e-01 6.17879468e-01 6.02403334e-01 5.86824089e-01
- 5.71157419e-01 5.55419100e-01 5.39624978e-01 5.23790958e-01
- 5.07932982e-01 4.92067018e-01 4.76209042e-01 4.60375022e-01
- 4.44580900e-01 4.28842581e-01 4.13175911e-01 3.97596666e-01
- 3.82120532e-01 3.66763093e-01 3.51539812e-01 3.36466018e-01
- 3.21556889e-01 3.06827437e-01 2.92292493e-01 2.77966694e-01
- 2.63864463e-01 2.50000000e-01 2.36387266e-01 2.23039968e-01
- 2.09971545e-01 1.97195156e-01 1.84723666e-01 1.72569633e-01
- 1.60745294e-01 1.49262556e-01 1.38132981e-01 1.27367775e-01
- 1.16977778e-01 1.06973453e-01 9.73648712e-02 8.81617093e-02
- 7.93732336e-02 7.10082934e-02 6.30753115e-02 5.55822757e-02
- 4.85367309e-02 4.19457713e-02 3.58160335e-02 3.01536896e-02
- 2.49644411e-02 2.02535132e-02 1.60256493e-02 1.22851066e-02
- 9.03565137e-03 6.28055566e-03 4.02259358e-03 2.26403871e-03
- 1.00666176e-03 2.51728808e-04] 
- [ 0.          0.00347789  0.00672602  0.00996571  0.01323204  0.01650328
-  0.01980878  0.02317628  0.02661133  0.03010495  0.03365667  0.03724499
-  0.04086488  0.04450757  0.04815953  0.05181257  0.05545525  0.0590815
-  0.06268667  0.06624826  0.06976721  0.07324567  0.07664624  0.07999416
-  0.08324662  0.08643356  0.08950536  0.09248551  0.09536027  0.09810166
-  0.10072646  0.10322516  0.105572    0.1077613   0.10979718  0.11166869
-  0.11336795  0.11488205  0.11619867  0.11730959  0.11819634  0.11884432
-  0.11921881  0.11927451  0.11898268  0.11836987  0.11748837  0.11635763
-  0.11501176  0.11345957  0.11170622  0.10976973  0.10766037  0.10536249
-  0.10292056  0.10034895  0.0976621   0.09485201  0.09192193  0.08891781
-  0.08585318  0.08270423  0.07950929  0.07629528  0.07304306  0.06977049
-  0.06651259  0.06324119  0.05999944  0.05679153  0.05361405  0.05050168
-  0.04743265  0.04444461  0.04152447  0.03869348  0.0359486   0.03329938
-  0.03075034  0.0283023   0.0259619   0.02373074  0.0216071   0.01953411
-  0.01756546  0.01571439  0.01395478  0.01230609  0.01078312  0.00938739
-  0.00809817  0.00692809  0.00589288  0.00499358  0.0042311   0.0036062
-  0.00311951  0.00277152  0.00256259  0.00249293  0.00252063  0.0026037
-  0.00274206  0.00293556  0.00318402  0.00348719  0.00384475  0.00425635
-  0.00472157  0.00516747  0.00548365  0.00582866  0.00620215  0.00654452
-  0.00666731  0.0067979   0.00692736  0.00662937  0.0063163   0.00590618
-  0.00529529  0.00465925  0.00379728  0.00288452  0.00179111  0.00060513
- -0.00073948 -0.00217554 -0.00376997 -0.00543477 -0.00726226 -0.00915913
- -0.01116457 -0.01326052 -0.0154248  -0.01767487 -0.01998183 -0.02233933
- -0.02474616 -0.02718491 -0.02964489 -0.03211674 -0.03457366 -0.03701392
- -0.0394263  -0.0417981  -0.04411643 -0.04635189 -0.04850595 -0.05056216
- -0.05250383 -0.05432106 -0.05599723 -0.05750118 -0.05879806 -0.05983815
- -0.06060047 -0.0611303  -0.06147551 -0.06165611 -0.06168998 -0.06159061
- -0.06136544 -0.06102167 -0.06056783 -0.06000196 -0.05933555 -0.05857795
- -0.05773654 -0.05680096 -0.05578404 -0.05469846 -0.05352939 -0.0522956
- -0.05099479 -0.04962894 -0.04820164 -0.04671799 -0.0451773  -0.04358869
- -0.0419391  -0.04023003 -0.03847292 -0.03666765 -0.03481456 -0.03290918
- -0.03095329 -0.02894203 -0.02686789 -0.02472747 -0.02250955 -0.02020322
- -0.01779699 -0.01526995 -0.01257333 -0.00961691 -0.0063974  -0.00312529] [3440000000.0, 41800000000.0, 27700000000.0, 13600000000.0, 200000000000.0, 205000000000.0, 118000000000.0, 98460000.0, 114500000000.0, 1000000.0, 4560000.0] [3440000000.0, 14000000000.0, 13650000000.0, 13300000000.0, 200000000000.0, 205000000000.0, 118000000000.0, 98460000.0, 8390000000.0, 1000000.0, 4560000.0] [1323000000.0, 2630000000.0, 7200000000.0, 11800000000.0, 79300000000.0, 80000000000.0, 47600000000.0, 98460000.0, 5990000000.0, 384615.3846153846, 1520000.0] [0.3, 0.28, 0.39, 0.49, 0.3, 0.3, 0.3, 0.3, 0.27, 0.3, 0.49] [1235.0, 1920.0, 1850.0, 1780.0, 7800.0, 7850.0, 7200.0, 200.0, 1220.0, 1150.0, 1100.0] [0.         0.15466451 0.64409767 0.92285544 1.        ] [3. 4. 3. 4.] [1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1.] [5.00000000e-05 2.82000000e-03 2.82000000e-03 5.00000000e-05
- 2.82000000e-03 4.44189253e-03 2.82000000e-03 5.00000000e-05
- 2.82000000e-03 2.82000000e-03 5.00000000e-05 2.82000000e-03
- 9.45861116e-06 2.82000000e-03] [0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0.] [1 3 3 1 3 9 3 1 3 3 1 3 2 3] [0.         0.15611071 0.6433223  0.9253126  1.        ] [3. 4. 3. 4.] [1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1. 1.] [5.00000000e-05 2.82000000e-03 2.82000000e-03 5.00000000e-05
- 2.82000000e-03 4.44189253e-03 2.82000000e-03 5.00000000e-05
- 2.82000000e-03 2.82000000e-03 5.00000000e-05 2.82000000e-03
- 9.45861116e-06 2.82000000e-03] [0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0. 0.] [1 3 3 1 3 9 3 1 3 3 1 3 2 3] 2 [0.20350056 0.59694928] [3. 3.] [1. 1. 1. 1. 1. 1.] [0.002 0.04  0.002 0.002 0.04  0.002] [0. 0. 0. 0. 0. 0.] [4 8 4 4 8 4]
-
- 
-29 1.419 0.0 0.0 0.4 [0.00000000e+00 2.51728808e-04 1.00666176e-03 2.26403871e-03
- 4.02259358e-03 6.28055566e-03 9.03565137e-03 1.22851066e-02
- 1.60256493e-02 2.02535132e-02 2.49644411e-02 3.01536896e-02
- 3.58160335e-02 4.19457713e-02 4.85367309e-02 5.55822757e-02
- 6.30753115e-02 7.10082934e-02 7.93732336e-02 8.81617093e-02
- 9.73648712e-02 1.06973453e-01 1.16977778e-01 1.27367775e-01
- 1.38132981e-01 1.49262556e-01 1.60745294e-01 1.72569633e-01
- 1.84723666e-01 1.97195156e-01 2.09971545e-01 2.23039968e-01
- 2.36387266e-01 2.50000000e-01 2.63864463e-01 2.77966694e-01
- 2.92292493e-01 3.06827437e-01 3.21556889e-01 3.36466018e-01
- 3.51539812e-01 3.66763093e-01 3.82120532e-01 3.97596666e-01
- 4.13175911e-01 4.28842581e-01 4.44580900e-01 4.60375022e-01
- 4.76209042e-01 4.92067018e-01 5.07932982e-01 5.23790958e-01
- 5.39624978e-01 5.55419100e-01 5.71157419e-01 5.86824089e-01
- 6.02403334e-01 6.17879468e-01 6.33236907e-01 6.48460188e-01
- 6.63533982e-01 6.78443111e-01 6.93172563e-01 7.07707507e-01
- 7.22033306e-01 7.36135537e-01 7.50000000e-01 7.63612734e-01
- 7.76960032e-01 7.90028455e-01 8.02804844e-01 8.15276334e-01
- 8.27430367e-01 8.39254706e-01 8.50737444e-01 8.61867019e-01
- 8.72632225e-01 8.83022222e-01 8.93026547e-01 9.02635129e-01
- 9.11838291e-01 9.20626766e-01 9.28991707e-01 9.36924689e-01
- 9.44417724e-01 9.51463269e-01 9.58054229e-01 9.64183967e-01
- 9.69846310e-01 9.75035559e-01 9.79746487e-01 9.83974351e-01
- 9.87714893e-01 9.90964349e-01 9.93719444e-01 9.95977406e-01
- 9.97735961e-01 9.98993338e-01 9.99748271e-01 1.00000000e+00
- 1.00000000e+00 9.99748271e-01 9.98993338e-01 9.97735961e-01
- 9.95977406e-01 9.93719444e-01 9.90964349e-01 9.87714893e-01
- 9.83974351e-01 9.79746487e-01 9.75035559e-01 9.69846310e-01
- 9.64183967e-01 9.58054229e-01 9.51463269e-01 9.44417724e-01
- 9.36924689e-01 9.28991707e-01 9.20626766e-01 9.11838291e-01
- 9.02635129e-01 8.93026547e-01 8.83022222e-01 8.72632225e-01
- 8.61867019e-01 8.50737444e-01 8.39254706e-01 8.27430367e-01
- 8.15276334e-01 8.02804844e-01 7.90028455e-01 7.76960032e-01
- 7.63612734e-01 7.50000000e-01 7.36135537e-01 7.22033306e-01
- 7.07707507e-01 6.93172563e-01 6.78443111e-01 6.63533982e-01
- 6.48460188e-01 6.33236907e-01 6.17879468e-01 6.02403334e-01
- 5.86824089e-01 5.71157419e-01 5.55419100e-01 5.39624978e-01
- 5.23790958e-01 5.07932982e-01 4.92067018e-01 4.76209042e-01
- 4.60375022e-01 4.44580900e-01 4.28842581e-01 4.13175911e-01
- 3.97596666e-01 3.82120532e-01 3.66763093e-01 3.51539812e-01
- 3.36466018e-01 3.21556889e-01 3.06827437e-01 2.92292493e-01
- 2.77966694e-01 2.63864463e-01 2.50000000e-01 2.36387266e-01
- 2.23039968e-01 2.09971545e-01 1.97195156e-01 1.84723666e-01
- 1.72569633e-01 1.60745294e-01 1.49262556e-01 1.38132981e-01
- 1.27367775e-01 1.16977778e-01 1.06973453e-01 9.73648712e-02
- 8.81617093e-02 7.93732336e-02 7.10082934e-02 6.30753115e-02
- 5.55822757e-02 4.85367309e-02 4.19457713e-02 3.58160335e-02
- 3.01536896e-02 2.49644411e-02 2.02535132e-02 1.60256493e-02
- 1.22851066e-02 9.03565137e-03 6.28055566e-03 4.02259358e-03
- 2.26403871e-03 1.00666176e-03 2.51728808e-04]  [ 0.          0.0034671   0.00671341  0.00995348  0.01321845  0.01648816
-  0.01979262  0.02315794  0.02659061  0.03008164  0.03363047  0.03721579
-  0.04083245  0.04447173  0.0481201   0.05176938  0.05540812  0.05903027
-  0.06263114  0.06618782  0.06970218  0.07317538  0.07657107  0.07991397
-  0.08316075  0.08634187  0.08940828  0.09238236  0.0952515   0.0979866
-  0.10060498  0.10309717  0.10543807  0.10762067  0.10964979  0.11151446
-  0.11320679  0.11471394  0.11602356  0.11712742  0.11800712  0.11864803
-  0.11901554  0.11906439  0.11876577  0.11814601  0.11725727  0.1161192
-  0.11476577  0.11320604  0.11144507  0.10950083  0.10738297  0.1050773
-  0.10262754  0.10004807  0.09735332  0.09453479  0.09159683  0.08858483
-  0.08551232  0.08235515  0.07915241  0.07593067  0.0726705   0.06939035
-  0.06612493  0.06284594  0.05959689  0.05638166  0.05319714  0.05007779
-  0.047002    0.04400736  0.04108077  0.03824354  0.0354926   0.03283755
-  0.03028288  0.02782944  0.02548386  0.02324776  0.02111942  0.01904204
-  0.01706917  0.01521412  0.01345082  0.01179867  0.01027249  0.00887383
-  0.00758196  0.00640948  0.00537214  0.004471    0.00370695  0.00308076
-  0.00259308  0.00224438  0.00203502  0.00196521  0.00196521  0.00199304
-  0.00207652  0.00221555  0.00240999  0.00265966  0.0029643   0.00332359
-  0.00373719  0.00420467  0.00465301  0.00497195  0.00531995  0.00569669
-  0.00604245  0.00616896  0.0063035   0.00643703  0.00614346  0.00583503
-  0.00542971  0.00482392  0.00419317  0.0033367   0.00242967  0.00134218
-  0.00016232 -0.00117595 -0.00260551 -0.00419323 -0.00585113 -0.00767159
- -0.00956123 -0.01155933 -0.01364776 -0.01580439 -0.01804671 -0.02034578
- -0.02269528 -0.02509404 -0.02752462 -0.02997638 -0.03243995 -0.0348885
- -0.03732039 -0.03972439 -0.04208781 -0.04439776 -0.0466247  -0.04877041
- -0.05081831 -0.05275171 -0.05456073 -0.05622879 -0.05772474 -0.0590138
- -0.06004622 -0.06080085 -0.06132295 -0.06166046 -0.06183342 -0.0618597
- -0.06175283 -0.06152024 -0.06116915 -0.06070809 -0.06013533 -0.05946193
- -0.05869746 -0.05784932 -0.05690736 -0.05588399 -0.05479213 -0.05361712
- -0.05237739 -0.05107098 -0.04969954 -0.04826702 -0.04677815 -0.04523261
- -0.04363941 -0.04198521 -0.04027197 -0.0385109  -0.03670191 -0.03484534
- -0.03293645 -0.03097756 -0.02896357 -0.026887   -0.02474437 -0.02252463
- -0.02021677 -0.01780932 -0.01528142 -0.01258441 -0.00962776 -0.00640903
- -0.0031376 ] [3440000000.0, 41800000000.0, 27700000000.0, 13600000000.0, 200000000000.0, 205000000000.0, 118000000000.0, 98460000.0, 114500000000.0, 1000000.0, 4560000.0] [3440000000.0, 14000000000.0, 13650000000.0, 13300000000.0, 200000000000.0, 205000000000.0, 118000000000.0, 98460000.0, 8390000000.0, 1000000.0, 4560000.0] [1323000000.0, 2630000000.0, 7200000000.0, 11800000000.0, 79300000000.0, 80000000000.0, 47600000000.0, 98460000.0, 5990000000.0, 384615.3846153846, 1520000.0] [0.3, 0.28, 0.39, 0.49, 0.3, 0.3, 0.3, 0.3, 0.27, 0.3, 0.49] [1235.0, 1920.0, 1850.0, 1780.0, 7800.0, 7850.0, 7200.0, 200.0, 1220.0, 1150.0, 1100.0] [0.         0.13642206 0.66212094 1.        ] [3. 4. 3.] [1. 1. 1. 1. 1. 1. 1. 1. 1. 1.] [5.00e-05 2.82e-03 2.82e-03 5.00e-05 2.82e-03 4.70e-04 2.82e-03 5.00e-05
- 2.82e-03 2.82e-03] [0. 0. 0. 0. 0. 0. 0. 0. 0. 0.] [1 3 3 1 3 9 3 1 3 3] [0.         0.13858834 0.66136763 1.        ] [3. 4. 3.] [1. 1. 1. 1. 1. 1. 1. 1. 1. 1.] [5.00e-05 2.82e-03 2.82e-03 5.00e-05 2.82e-03 4.70e-04 2.82e-03 5.00e-05
- 2.82e-03 2.82e-03] [0. 0. 0. 0. 0. 0. 0. 0. 0. 0.] [1 3 3 1 3 9 3 1 3 3] 3 [0.18879231 0.61154576 1.        ] [3. 3. 3.] [1. 1. 1. 1. 1. 1. 1. 1. 1.] [2.00e-03 4.00e-02 2.00e-03 2.00e-03 4.00e-02 2.00e-03 5.00e-05 2.82e-03
- 2.82e-03] [0. 0. 0. 0. 0. 0. 0. 0. 0.] [4 8 4 4 8 4 1 3 3]
-
-0 (27308516755.423786, 27110066507.9137, 13952243195.444218, 18118671160.42156, 92074585.33274896, -15460.740430957281, -280289306.6835994, -0.0, 0.0, 0.0, -0.0013892862524830067, -6.888959056499682e-05, -0.0013884329481598547, 0.015400749667299764, 1276.1536419953366, 17.366065786576875, 1872.3066046676142, 1851.0336153527403, -2.5971291329272783, -0.0013858771158858443, 0.056855449043146725)
-1 (16652335849.561424, 16144645307.711245, 8034367444.044591, 10708928148.30381, 569789795.8471074, -2016668.8170878347, -272294671.45954186, -0.0, 0.0, 0.0, -0.0026410715610530188, 0.03623157968874161, -0.0024527549723935595, 0.061658463471006106, 774.8339755205895, 17.503484286448916, 1138.9517314248762, 1111.0694342328163, -19.84049447807963, -0.013104716954923214, 0.14440089527749475)
-2 (18473117022.61538, 13405078510.418785, 5045804154.236201, 10249700371.107042, 5194029271.970429, -23299060.664214015, -235755949.53862202, -0.0, 0.0, 0.0, -0.020334717941031043, 0.0659938233879191, -0.018061572403051035, 0.08899507620533767, 542.1354397205109, 17.398834933111498, 683.1069325506609, 862.894946848888, 40.954838306861106, 0.0003862522651320935, 0.22483245540118635)
-3 (16729123005.373446, 9442950078.618265, 2283145364.6513453, 10397748336.982285, 6869704905.300286, -48729536.7656376, -158208767.6642635, -0.0, 0.0, 0.0, -0.0038971456893817304, 0.033658056808765095, 0.0007894013577629268, 0.048873732600951084, 416.16320606314383, 15.854490997398763, 339.33238489255484, 609.7480553712666, 38.923305233426234, 0.06912487577291207, 0.3211365003752893)
-4 (10348388861.574236, 7080868983.96506, 1291944621.0352225, 9395936020.450193, 4946375901.665741, -42216488.4583947, -85058171.42358719, -0.0, 0.0, 0.0, 0.08779775314730143, 0.06763767277842916, 0.09229081146171639, 0.07669032822845125, 368.2146573796386, 14.508935319620239, 187.98546475963713, 515.2327403734753, 26.806639360156403, 0.23385610725183428, 0.37545404030257845)
-5 (7275716474.565289, 6425140486.535936, 973420039.9171844, 9078896032.773207, 3953860391.249667, -28693204.02328584, -49861703.650871664, -0.0, 0.0, 0.0, 0.13639709742106, 0.11614733364389941, 0.13955752616919284, 0.12163937843336277, 340.4047970813398, 13.837220430643518, 118.67842445351874, 484.66408070087874, 22.881368653398653, 0.3052283751353755, 0.4766003182499454)
-6 (6194875654.490169, 6196377070.625498, 897289593.4256071, 9126064020.618538, 3397508576.42813, -23060702.748218335, -34537770.20368287, -0.0, 0.0, 0.0, 0.15919829491568802, 0.1307053715946589, 0.16172520056703094, 0.13448989141893927, 336.2777697831578, 13.52483186027556, 100.74910297313997, 478.1048754478975, 20.65051273756747, 0.3287282123266766, 0.49070814959996417)
-7 (5433488053.343419, 5955692384.1130495, 820895994.6497997, 9167035826.157984, 2939374751.0531464, -18550567.468462713, -24241173.982702494, -0.0, 0.0, 0.0, 0.1780688215045793, 0.13727837461216816, 0.18009243827888372, 0.13992275982483754, 332.1819061188258, 13.24929321856374, 89.14833545374321, 463.6480183376485, 18.857528076737985, 0.33833436596314487, 0.4953330605677446)
-8 (4653520688.226214, 5488438191.599841, 713834694.2130578, 9148766184.863085, 2504887544.6905937, -14695771.035853803, -17237052.73731969, -0.0, 0.0, 0.0, 0.18349332679401217, 0.13538659064317682, 0.18509963858938502, 0.13727067559944545, 322.94743030818853, 12.751897732112507, 76.00553473120021, 424.29194837828925, 17.64739173946489, 0.33093436381554375, 0.48333135178322356)
-9 (4004737840.8732615, 5032344198.254628, 607892470.2696717, 9116671504.624172, 2134725079.1916945, -11669381.401902337, -12121354.720229466, -0.0, 0.0, 0.0, 0.17355697351436136, 0.13145737891136838, 0.17483697810770263, 0.1327869600763891, 313.4922178427712, 12.258961237823495, 65.18880266018422, 385.0680942368176, 16.57718529436479, 0.308205100445597, 0.4679159768157717)
-10 (3477025033.4429846, 4656889263.946102, 522747979.8100182, 9085103450.281673, 1832460273.3449588, -9212250.772720564, -8181402.19665206, -0.0, 0.0, 0.0, 0.16725901822923883, 0.12741178901958058, 0.16827301337229192, 0.12831231835768056, 305.1420941669453, 11.82512406916226, 56.49716544938471, 352.53131975129963, 15.550379658005031, 0.2894833285062009, 0.4540071782171048)
-11 (2967372602.7120414, 4281981679.9274898, 440012888.9614443, 9039012721.861242, 1540711709.1013203, -6855291.547289756, -4798173.998253582, -0.0, 0.0, 0.0, 0.16687495266488314, 0.12176736338564619, 0.16763336420364666, 0.12229819281867144, 289.20956446099444, 11.36781560085032, 47.81042922781151, 314.105308504208, 14.396230787357796, 0.27065194571286266, 0.42111059818711744)
-12 (2425174220.2446566, 3865248717.9758086, 360351670.7001954, 8915864611.529379, 1237412449.7844906, -4395984.10475484, -2116124.4244569466, -0.0, 0.0, 0.0, 0.17714663427499983, 0.11289719108630905, 0.17763968618822318, 0.11313453480076166, 264.786979053903, 10.869776382429956, 38.70881174703595, 269.0323616633825, 13.036242434232586, 0.2549110775022159, 0.36433458445413536)
-13 (1934971209.1790102, 3438031362.097637, 295049666.73697877, 8712833880.018932, 961135870.0729796, -2531160.5100712692, -252617.2731182836, -0.0, 0.0, 0.0, 0.18763861150744035, 0.10315617473774545, 0.1879291209495951, 0.10318516843561654, 253.78623170516514, 10.368894892095476, 31.280173472294038, 237.7690969467521, 11.624335907246959, 0.2479218272779372, 0.34559386923067864)
-14 (1550224031.085044, 3009111460.853723, 238188173.82161444, 8435410831.982225, 733870634.9989922, -1542013.2336249878, 916447.5829342771, -0.0, 0.0, 0.0, 0.18723440744897782, 0.09312019423802369, 0.18741720983576596, 0.09301155132765593, 241.96389521358356, 9.885384900716382, 25.44510955555225, 207.705174782151, 10.293155634078294, 0.23471591842508058, 0.3250462683784696)
-15 (1243850945.639248, 2611321823.5500717, 191749294.19840983, 8105116320.101374, 555444451.6842866, -1054594.3072677967, 1626756.6342297513, -0.0, 0.0, 0.0, 0.1824982661388704, 0.08294492626558049, 0.18262838078073584, 0.0827442188889084, 229.85266073990528, 9.43091287265537, 20.744737808149992, 180.8228710239727, 9.071895232803094, 0.22030906302446704, 0.30380811731638)
-16 (993936577.6806138, 2250650282.571106, 153603148.64042357, 7736964882.957831, 415550900.2204178, -738060.3791703708, 2007362.2330203503, -0.0, 0.0, 0.0, 0.17675216272796368, 0.07311531690289981, 0.17684755677334651, 0.07285586603768496, 217.65390191975538, 9.001970630866321, 16.856788533528785, 156.96576944662831, 7.942726190939896, 0.2061646834945001, 0.282744742321804)
-17 (787710875.3640277, 1923876794.7291553, 121222942.80622464, 7341645894.403406, 304437473.53028923, -489501.24417329556, 2157812.9882384557, -0.0, 0.0, 0.0, 0.16985210486261923, 0.06403567354139655, 0.1699187794545587, 0.06374175947644779, 205.31421686306803, 8.582995643112678, 13.585993265134803, 135.38323715046857, 6.86619501560346, 0.19142054180967147, 0.26220195366953036)
-18 (612440265.2301638, 1635344171.5796325, 94667880.16715504, 6932724532.942047, 216088703.75916198, -281803.21641578025, 2136074.176126303, -0.0, 0.0, 0.0, 0.16683570688513746, 0.056102323449035214, 0.16687635514904536, 0.05579420877262244, 192.92737767921224, 8.164625691428792, 10.73857534518256, 116.06724353876612, 5.8028777053435245, 0.1808309259994654, 0.24310987579727295)
-19 (459538315.7911478, 1382661257.8225594, 69457587.44386901, 6516937256.484205, 147680152.05942053, -171734.0256130834, 1886624.3394836634, -0.0, 0.0, 0.0, 0.1588563986859967, 0.048790383933755496, 0.15888275063920024, 0.048500888383126016, 167.55512409668037, 7.7343093103547, 8.050683934825706, 91.36259305490424, 4.789260729505307, 0.16238788061956982, 0.17389187639084008)
-20 (357106704.24688035, 1179560675.3553565, 53071762.56827737, 6108667346.997217, 102751275.20285574, -123216.64818663872, 1685814.1683255453, -0.0, 0.0, 0.0, 0.1498960046559266, 0.044016637301175054, 0.1499161754462459, 0.0437406664432834, 152.49815583271507, 7.349615532982778, 6.308405239671117, 75.38230484210744, 3.991209985383262, 0.14853679897388872, 0.13941748124412578)
-21 (299016635.6225521, 1021610521.0893968, 43120434.712814726, 5667661921.961301, 76516960.90512578, -93945.08048080327, 1705807.5405396102, -0.0, 0.0, 0.0, 0.1344725508397692, 0.04215785170147005, 0.13448912647215064, 0.04185687970207646, 143.2207248573657, 7.031150181885307, 5.373077530331589, 65.13930614169062, 3.4851017836745113, 0.1320128793545904, 0.13023247386479678)
-22 (247637198.3259574, 891602616.7664934, 35393961.96708839, 5185278210.324141, 56781686.04964811, -69838.22540674981, 1822461.845561603, -0.0, 0.0, 0.0, 0.12057884587671626, 0.04250332580013029, 0.12059231443569689, 0.04215185732117134, 133.88529890703276, 6.717769666380972, 4.546080790034097, 56.57246659115583, 3.0318021726441176, 0.1174004521501232, 0.12508903090959456)
-23 (200426527.77546898, 770944442.5644885, 28851958.485642016, 4657369173.802529, 40421267.45911137, -53548.7053885457, 1945454.6476114895, -0.0, 0.0, 0.0, 0.10719043444017132, 0.04349931897704033, 0.10720193206971071, 0.04308160361563961, 124.23327484596352, 6.4048673057884, 3.7870283903973174, 48.97243337370113, 2.567492823486248, 0.10339860805108654, 0.12188566907812229)
-24 (157612891.3791863, 659067108.24051, 23365663.857873134, 4087634651.835455, 27349927.35892557, -43614.64585858729, 2072373.0477375882, -0.0, 0.0, 0.0, 0.09448263018555397, 0.045360663732591296, 0.09449330008381383, 0.04485367785340881, 114.28192855605849, 6.092092741416388, 3.097373088877653, 42.222710638109135, 2.106563557501107, 0.0901718769818779, 0.120655730311121)
-25 (119804240.14273137, 557896539.7967145, 19067306.57811477, 3481314087.11242, 17189413.171736002, -43065.19264158094, 2242370.566114801, -0.0, 0.0, 0.0, 0.08499964952795365, 0.048782324854382546, 0.08501201991211434, 0.0481382087217946, 104.21437303474272, 5.790547649355169, 2.490968037601977, 36.427084898000416, 1.643468666633411, 0.08023895064464297, 0.12188453443990738)
-26 (85104226.73499094, 457709468.3924512, 14995498.650817856, 2835923510.9259586, 9529163.84392222, -49166.93926452802, 2435377.5764779383, -0.0, 0.0, 0.0, 0.07304030739309059, 0.05404577277180489, 0.0730576445811618, 0.05318701280627861, 93.41951343669862, 5.45748012414415, 1.9186482584405145, 30.75335698435294, 1.1946270676589135, 0.0679598981347642, 0.124607198754151)
-27 (52694003.10580707, 347984076.1403547, 10966320.934678862, 2144566199.4409988, 4207305.858765096, -59143.14964645324, 2614402.6181055317, -0.0, 0.0, 0.0, 0.06064663277997576, 0.06159859895216549, 0.060674210921262274, 0.06037951668798218, 80.85410866798019, 5.0210785677790994, 1.3418531919160621, 24.274760539890895, 0.781610281493708, 0.05555809635123684, 0.12691797013211387)
-28 (15065999.21600573, 142438583.03857857, 4144759.8663949445, 1281868072.4096262, 1249705.9314852888, 9853.55631242007, 1709152.6476751757, -0.0, 0.0, 0.0, 0.03989537248173026, 0.05410078479942968, 0.03988768560932544, 0.052767455196748844, 47.9175130819107, 3.5952347925532355, 0.4247633425395061, 7.83116212575791, 0.4638045205037931, 0.03689370645715227, 0.09401330692297125)
-29 (4527565.186165874, 87084882.29444353, 2848798.552905994, 562865586.8124115, 562874.4137147996, -56457.08955084603, 2870898.5415248065, -0.0, 0.0, 0.0, 0.03316932104900284, 0.1096869524349319, 0.033269624018083914, 0.10458644778730702, 37.69127664825495, 3.3331598210278766, 0.26323750944008006, 6.045513520847946, 0.4466784247731151, 0.03278434630655923, 0.10327104335910127)
-        '''
-
-
-if __name__ == "__main__":
-    unittest.main()
